@@ -20,28 +20,28 @@
   "dependencies": {
     "@frequency-chain/api-augment": "file:../js/api-augment/dist/frequency-chain-api-augment-0.0.0.tgz",
     "@helia/unixfs": "^3.0.7",
-    "@noble/curves": "^1.5.0",
-    "@polkadot/api": "12.3.1",
-    "@polkadot/types": "12.3.1",
-    "@polkadot/util": "13.0.2",
-    "helia": "^4.2.5",
-    "multiformats": "^13.2.2",
+    "@noble/curves": "^1.6.0",
+    "@polkadot/api": "13.0.1",
+    "@polkadot/types": "13.0.1",
+    "@polkadot/util": "13.1.1",
+    "helia": "^4.2.6",
+    "multiformats": "^13.3.0",
     "rxjs": "^7.8.1",
     "workerpool": "^9.1.3"
   },
   "devDependencies": {
-    "@eslint/js": "^9.9.0",
-    "@types/mocha": "^10.0.7",
+    "@eslint/js": "^9.10.0",
+    "@types/mocha": "^10.0.8",
     "@types/workerpool": "^6.4.7",
-    "eslint": "^9.9.0",
+    "eslint": "^9.10.0",
     "eslint-plugin-mocha": "^10.5.0",
     "globals": "^15.9.0",
     "mocha": "^10.7.3",
     "node-datachannel": "^0.11.0",
     "prettier": "^3.3.3",
-    "sinon": "^18.0.0",
-    "tsx": "^4.17.0",
-    "typescript": "^5.5.4",
-    "typescript-eslint": "^8.1.0"
+    "sinon": "^19.0.2",
+    "tsx": "^4.19.1",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0"
   }
 }
