@@ -33,21 +33,21 @@
   "author": "frequency-chain",
   "license": "Apache-2.0",
   "dependencies": {
-    "@polkadot/api": "^12.3.1",
-    "@polkadot/rpc-provider": "^12.3.1",
-    "@polkadot/types": "^12.3.1",
+    "@polkadot/api": "^13.0.1",
+    "@polkadot/rpc-provider": "^13.0.1",
+    "@polkadot/types": "^13.0.1",
     "globals": "^15.9.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.9.0",
-    "@polkadot/typegen": "^12.3.1",
-    "@types/mocha": "^10.0.7",
-    "eslint": "^9.9.0",
+    "@eslint/js": "^9.10.0",
+    "@polkadot/typegen": "^13.0.1",
+    "@types/mocha": "^10.0.8",
+    "eslint": "^9.10.0",
     "eslint-plugin-mocha": "^10.5.0",
     "mocha": "10.7.3",
     "prettier": "^3.3.3",
-    "tsx": "^4.17.0",
-    "typescript": "^5.5.4",
-    "typescript-eslint": "^8.1.0"
+    "tsx": "^4.19.1",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0"
   }
 }
