@@ -175,9 +175,11 @@ pub fn inherent_benchmark_data() -> Result<InherentData> {
 	let timestamp = sp_timestamp::InherentDataProvider::new(d.into());
 	let mock_para_inherent_provider =
 		cumulus_client_parachain_inherent::MockValidationDataInherentDataProvider {
+			para_id: 1000.into(),
+			current_para_block_head: Some(cumulus_primitives_core::relay_chain::HeadData::default()),
 			current_para_block: 0,
 			relay_offset: 1,
-			relay_blocks_per_para_block: 2,
+			relay_blocks_per_para_block: 1,
 			xcm_config: Default::default(),
 			raw_downward_messages: Default::default(),
 			raw_horizontal_messages: Default::default(),
