@@ -12,12 +12,11 @@ use sp_runtime::traits::AccountIdConversion;
 use super::{get_account_id_from_seed, get_collator_keys_from_seed, get_properties, Extensions};
 
 /// Specialized `ChainSpec` for the normal parachain runtime.
-pub type ChainSpec =
-	sc_service::GenericChainSpec<frequency_runtime::RuntimeGenesisConfig, Extensions>;
+pub type ChainSpec = sc_service::GenericChainSpec<Extensions>;
 use sp_core::sr25519;
 
 // Generic chain spec, in case when we don't have the native runtime.
-pub type RelayChainSpec = sc_service::GenericChainSpec<(), RelayChainExtensions>;
+pub type RelayChainSpec = sc_service::GenericChainSpec<RelayChainExtensions>;
 
 #[allow(clippy::unwrap_used)]
 /// Generates the Frequency Paseo chain spec from the raw json
