@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_capacity`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -61,8 +61,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `174`
 		//  Estimated: `5259`
-		// Minimum execution time: 37_829_000 picoseconds.
-		Weight::from_parts(38_703_000, 5259)
+		// Minimum execution time: 37_795_000 picoseconds.
+		Weight::from_parts(38_834_000, 5259)
 			.saturating_add(T::DbWeight::get().reads(7_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -78,8 +78,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `285`
 		//  Estimated: `5259`
-		// Minimum execution time: 25_467_000 picoseconds.
-		Weight::from_parts(26_355_000, 5259)
+		// Minimum execution time: 25_033_000 picoseconds.
+		Weight::from_parts(25_536_000, 5259)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -91,8 +91,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `1984`
-		// Minimum execution time: 2_198_000 picoseconds.
-		Weight::from_parts(2_311_000, 1984)
+		// Minimum execution time: 2_221_000 picoseconds.
+		Weight::from_parts(2_294_000, 1984)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -108,8 +108,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `271`
 		//  Estimated: `4081`
-		// Minimum execution time: 25_269_000 picoseconds.
-		Weight::from_parts(26_226_000, 4081)
+		// Minimum execution time: 24_068_000 picoseconds.
+		Weight::from_parts(24_809_000, 4081)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -119,8 +119,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 4_325_000 picoseconds.
-		Weight::from_parts(7_213_000, 0)
+		// Minimum execution time: 3_840_000 picoseconds.
+		Weight::from_parts(4_067_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 }
@@ -145,8 +145,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `174`
 		//  Estimated: `5259`
-		// Minimum execution time: 37_829_000 picoseconds.
-		Weight::from_parts(38_703_000, 5259)
+		// Minimum execution time: 37_795_000 picoseconds.
+		Weight::from_parts(38_834_000, 5259)
 			.saturating_add(RocksDbWeight::get().reads(7_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -162,8 +162,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `285`
 		//  Estimated: `5259`
-		// Minimum execution time: 25_467_000 picoseconds.
-		Weight::from_parts(26_355_000, 5259)
+		// Minimum execution time: 25_033_000 picoseconds.
+		Weight::from_parts(25_536_000, 5259)
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
@@ -175,8 +175,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `1984`
-		// Minimum execution time: 2_198_000 picoseconds.
-		Weight::from_parts(2_311_000, 1984)
+		// Minimum execution time: 2_221_000 picoseconds.
+		Weight::from_parts(2_294_000, 1984)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -192,8 +192,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `271`
 		//  Estimated: `4081`
-		// Minimum execution time: 25_269_000 picoseconds.
-		Weight::from_parts(26_226_000, 4081)
+		// Minimum execution time: 24_068_000 picoseconds.
+		Weight::from_parts(24_809_000, 4081)
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -203,8 +203,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 4_325_000 picoseconds.
-		Weight::from_parts(7_213_000, 0)
+		// Minimum execution time: 3_840_000 picoseconds.
+		Weight::from_parts(4_067_000, 0)
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 }
@@ -216,6 +216,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
