@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_frequency_tx_payment`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -44,18 +44,18 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_192_000 picoseconds.
-		Weight::from_parts(3_333_000, 0)
+		// Minimum execution time: 3_096_000 picoseconds.
+		Weight::from_parts(3_267_000, 0)
 	}
 	/// The range of component `n` is `[0, 10]`.
 	fn pay_with_capacity_batch_all(n: u32, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 5_242_000 picoseconds.
-		Weight::from_parts(6_274_375, 0)
-			// Standard Error: 11_416
-			.saturating_add(Weight::from_parts(3_172_765, 0).saturating_mul(n.into()))
+		// Minimum execution time: 5_018_000 picoseconds.
+		Weight::from_parts(5_849_351, 0)
+			// Standard Error: 9_305
+			.saturating_add(Weight::from_parts(3_104_086, 0).saturating_mul(n.into()))
 	}
 }
 
@@ -65,18 +65,18 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_192_000 picoseconds.
-		Weight::from_parts(3_333_000, 0)
+		// Minimum execution time: 3_096_000 picoseconds.
+		Weight::from_parts(3_267_000, 0)
 	}
 	/// The range of component `n` is `[0, 10]`.
 	fn pay_with_capacity_batch_all(n: u32, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 5_242_000 picoseconds.
-		Weight::from_parts(6_274_375, 0)
-			// Standard Error: 11_416
-			.saturating_add(Weight::from_parts(3_172_765, 0).saturating_mul(n.into()))
+		// Minimum execution time: 5_018_000 picoseconds.
+		Weight::from_parts(5_849_351, 0)
+			// Standard Error: 9_305
+			.saturating_add(Weight::from_parts(3_104_086, 0).saturating_mul(n.into()))
 	}
 }
 
@@ -87,6 +87,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
