@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_handles`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -54,10 +54,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `116`
 		//  Estimated: `4019`
-		// Minimum execution time: 63_905_000 picoseconds.
-		Weight::from_parts(65_360_796, 4019)
-			// Standard Error: 8_534
-			.saturating_add(Weight::from_parts(70_441, 0).saturating_mul(b.into()))
+		// Minimum execution time: 68_677_000 picoseconds.
+		Weight::from_parts(70_300_113, 4019)
+			// Standard Error: 9_562
+			.saturating_add(Weight::from_parts(85_735, 0).saturating_mul(b.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -74,10 +74,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `222 + b * (1 ±0)`
 		//  Estimated: `4019`
-		// Minimum execution time: 71_561_000 picoseconds.
-		Weight::from_parts(72_934_394, 4019)
-			// Standard Error: 8_167
-			.saturating_add(Weight::from_parts(149_513, 0).saturating_mul(b.into()))
+		// Minimum execution time: 77_227_000 picoseconds.
+		Weight::from_parts(78_870_608, 4019)
+			// Standard Error: 7_816
+			.saturating_add(Weight::from_parts(170_182, 0).saturating_mul(b.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -91,8 +91,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `231`
 		//  Estimated: `4019`
-		// Minimum execution time: 18_701_000 picoseconds.
-		Weight::from_parts(19_862_000, 4019)
+		// Minimum execution time: 19_290_000 picoseconds.
+		Weight::from_parts(20_118_000, 4019)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -113,10 +113,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `116`
 		//  Estimated: `4019`
-		// Minimum execution time: 63_905_000 picoseconds.
-		Weight::from_parts(65_360_796, 4019)
-			// Standard Error: 8_534
-			.saturating_add(Weight::from_parts(70_441, 0).saturating_mul(b.into()))
+		// Minimum execution time: 68_677_000 picoseconds.
+		Weight::from_parts(70_300_113, 4019)
+			// Standard Error: 9_562
+			.saturating_add(Weight::from_parts(85_735, 0).saturating_mul(b.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -133,10 +133,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `222 + b * (1 ±0)`
 		//  Estimated: `4019`
-		// Minimum execution time: 71_561_000 picoseconds.
-		Weight::from_parts(72_934_394, 4019)
-			// Standard Error: 8_167
-			.saturating_add(Weight::from_parts(149_513, 0).saturating_mul(b.into()))
+		// Minimum execution time: 77_227_000 picoseconds.
+		Weight::from_parts(78_870_608, 4019)
+			// Standard Error: 7_816
+			.saturating_add(Weight::from_parts(170_182, 0).saturating_mul(b.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -150,8 +150,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `231`
 		//  Estimated: `4019`
-		// Minimum execution time: 18_701_000 picoseconds.
-		Weight::from_parts(19_862_000, 4019)
+		// Minimum execution time: 19_290_000 picoseconds.
+		Weight::from_parts(20_118_000, 4019)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
@@ -164,6 +164,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
