@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_messages`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -53,10 +53,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `329`
 		//  Estimated: `4177`
-		// Minimum execution time: 27_775_000 picoseconds.
-		Weight::from_parts(29_153_272, 4177)
-			// Standard Error: 65
-			.saturating_add(Weight::from_parts(925, 0).saturating_mul(n.into()))
+		// Minimum execution time: 27_851_000 picoseconds.
+		Weight::from_parts(28_914_770, 4177)
+			// Standard Error: 34
+			.saturating_add(Weight::from_parts(938, 0).saturating_mul(n.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -70,8 +70,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `722`
 		//  Estimated: `4008`
-		// Minimum execution time: 28_366_000 picoseconds.
-		Weight::from_parts(28_951_000, 4008)
+		// Minimum execution time: 28_198_000 picoseconds.
+		Weight::from_parts(28_689_000, 4008)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -92,10 +92,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `329`
 		//  Estimated: `4177`
-		// Minimum execution time: 27_775_000 picoseconds.
-		Weight::from_parts(29_153_272, 4177)
-			// Standard Error: 65
-			.saturating_add(Weight::from_parts(925, 0).saturating_mul(n.into()))
+		// Minimum execution time: 27_851_000 picoseconds.
+		Weight::from_parts(28_914_770, 4177)
+			// Standard Error: 34
+			.saturating_add(Weight::from_parts(938, 0).saturating_mul(n.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -109,8 +109,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `722`
 		//  Estimated: `4008`
-		// Minimum execution time: 28_366_000 picoseconds.
-		Weight::from_parts(28_951_000, 4008)
+		// Minimum execution time: 28_198_000 picoseconds.
+		Weight::from_parts(28_689_000, 4008)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -123,6 +123,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
