@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_msa`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -60,8 +60,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `4008`
-		// Minimum execution time: 11_983_000 picoseconds.
-		Weight::from_parts(12_300_000, 4008)
+		// Minimum execution time: 11_382_000 picoseconds.
+		Weight::from_parts(12_192_000, 4008)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -86,10 +86,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `1279`
 		//  Estimated: `6531`
-		// Minimum execution time: 98_502_000 picoseconds.
-		Weight::from_parts(101_717_762, 6531)
-			// Standard Error: 20_447
-			.saturating_add(Weight::from_parts(115_144, 0).saturating_mul(s.into()))
+		// Minimum execution time: 97_628_000 picoseconds.
+		Weight::from_parts(101_100_510, 6531)
+			// Standard Error: 10_151
+			.saturating_add(Weight::from_parts(93_683, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(10_u64))
 			.saturating_add(T::DbWeight::get().writes(7_u64))
 	}
@@ -101,8 +101,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161`
 		//  Estimated: `4177`
-		// Minimum execution time: 13_939_000 picoseconds.
-		Weight::from_parts(14_243_000, 4177)
+		// Minimum execution time: 12_968_000 picoseconds.
+		Weight::from_parts(13_604_000, 4177)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -116,10 +116,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	/// Proof: `Msa::PublicKeyCountForMsaId` (`max_values`: None, `max_size`: Some(17), added: 2492, mode: `MaxEncodedLen`)
 	fn add_public_key_to_msa() -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `1574`
+		//  Measured:  `1578`
 		//  Estimated: `9981`
-		// Minimum execution time: 142_408_000 picoseconds.
-		Weight::from_parts(143_966_000, 9981)
+		// Minimum execution time: 147_619_000 picoseconds.
+		Weight::from_parts(151_805_000, 9981)
 			.saturating_add(T::DbWeight::get().reads(8_u64))
 			.saturating_add(T::DbWeight::get().writes(7_u64))
 	}
@@ -131,8 +131,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `254`
 		//  Estimated: `6531`
-		// Minimum execution time: 24_633_000 picoseconds.
-		Weight::from_parts(25_527_000, 6531)
+		// Minimum execution time: 23_557_000 picoseconds.
+		Weight::from_parts(24_636_000, 6531)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -144,8 +144,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `71`
 		//  Estimated: `4008`
-		// Minimum execution time: 18_092_000 picoseconds.
-		Weight::from_parts(18_435_000, 4008)
+		// Minimum execution time: 17_191_000 picoseconds.
+		Weight::from_parts(17_633_000, 4008)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -164,12 +164,12 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	/// The range of component `s` is `[0, 30]`.
 	fn grant_delegation(s: u32, ) -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `1331`
+		//  Measured:  `1327`
 		//  Estimated: `6531`
-		// Minimum execution time: 88_976_000 picoseconds.
-		Weight::from_parts(93_844_651, 6531)
-			// Standard Error: 26_835
-			.saturating_add(Weight::from_parts(157_003, 0).saturating_mul(s.into()))
+		// Minimum execution time: 89_674_000 picoseconds.
+		Weight::from_parts(92_195_302, 6531)
+			// Standard Error: 13_523
+			.saturating_add(Weight::from_parts(121_658, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(8_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -181,8 +181,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161`
 		//  Estimated: `4177`
-		// Minimum execution time: 13_962_000 picoseconds.
-		Weight::from_parts(14_566_000, 4177)
+		// Minimum execution time: 13_045_000 picoseconds.
+		Weight::from_parts(13_415_000, 4177)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -194,8 +194,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `67`
 		//  Estimated: `4008`
-		// Minimum execution time: 10_804_000 picoseconds.
-		Weight::from_parts(11_092_000, 4008)
+		// Minimum execution time: 10_533_000 picoseconds.
+		Weight::from_parts(10_914_000, 4008)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -207,8 +207,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `67`
 		//  Estimated: `4008`
-		// Minimum execution time: 10_857_000 picoseconds.
-		Weight::from_parts(11_268_000, 4008)
+		// Minimum execution time: 10_301_000 picoseconds.
+		Weight::from_parts(10_646_000, 4008)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -226,8 +226,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `147`
 		//  Estimated: `4107`
-		// Minimum execution time: 19_016_000 picoseconds.
-		Weight::from_parts(19_618_000, 4107)
+		// Minimum execution time: 18_100_000 picoseconds.
+		Weight::from_parts(18_367_000, 4107)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -242,10 +242,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `316 + s * (6 ±0)`
 		//  Estimated: `4177`
-		// Minimum execution time: 16_854_000 picoseconds.
-		Weight::from_parts(17_824_532, 4177)
-			// Standard Error: 3_956
-			.saturating_add(Weight::from_parts(111_331, 0).saturating_mul(s.into()))
+		// Minimum execution time: 16_195_000 picoseconds.
+		Weight::from_parts(17_118_368, 4177)
+			// Standard Error: 4_262
+			.saturating_add(Weight::from_parts(108_183, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -263,8 +263,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `4008`
-		// Minimum execution time: 11_983_000 picoseconds.
-		Weight::from_parts(12_300_000, 4008)
+		// Minimum execution time: 11_382_000 picoseconds.
+		Weight::from_parts(12_192_000, 4008)
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -289,10 +289,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `1279`
 		//  Estimated: `6531`
-		// Minimum execution time: 98_502_000 picoseconds.
-		Weight::from_parts(101_717_762, 6531)
-			// Standard Error: 20_447
-			.saturating_add(Weight::from_parts(115_144, 0).saturating_mul(s.into()))
+		// Minimum execution time: 97_628_000 picoseconds.
+		Weight::from_parts(101_100_510, 6531)
+			// Standard Error: 10_151
+			.saturating_add(Weight::from_parts(93_683, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(10_u64))
 			.saturating_add(RocksDbWeight::get().writes(7_u64))
 	}
@@ -304,8 +304,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `161`
 		//  Estimated: `4177`
-		// Minimum execution time: 13_939_000 picoseconds.
-		Weight::from_parts(14_243_000, 4177)
+		// Minimum execution time: 12_968_000 picoseconds.
+		Weight::from_parts(13_604_000, 4177)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -319,10 +319,10 @@ impl WeightInfo for () {
 	/// Proof: `Msa::PublicKeyCountForMsaId` (`max_values`: None, `max_size`: Some(17), added: 2492, mode: `MaxEncodedLen`)
 	fn add_public_key_to_msa() -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `1574`
+		//  Measured:  `1578`
 		//  Estimated: `9981`
-		// Minimum execution time: 142_408_000 picoseconds.
-		Weight::from_parts(143_966_000, 9981)
+		// Minimum execution time: 147_619_000 picoseconds.
+		Weight::from_parts(151_805_000, 9981)
 			.saturating_add(RocksDbWeight::get().reads(8_u64))
 			.saturating_add(RocksDbWeight::get().writes(7_u64))
 	}
@@ -334,8 +334,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `254`
 		//  Estimated: `6531`
-		// Minimum execution time: 24_633_000 picoseconds.
-		Weight::from_parts(25_527_000, 6531)
+		// Minimum execution time: 23_557_000 picoseconds.
+		Weight::from_parts(24_636_000, 6531)
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
@@ -347,8 +347,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `71`
 		//  Estimated: `4008`
-		// Minimum execution time: 18_092_000 picoseconds.
-		Weight::from_parts(18_435_000, 4008)
+		// Minimum execution time: 17_191_000 picoseconds.
+		Weight::from_parts(17_633_000, 4008)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
@@ -367,12 +367,12 @@ impl WeightInfo for () {
 	/// The range of component `s` is `[0, 30]`.
 	fn grant_delegation(s: u32, ) -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `1331`
+		//  Measured:  `1327`
 		//  Estimated: `6531`
-		// Minimum execution time: 88_976_000 picoseconds.
-		Weight::from_parts(93_844_651, 6531)
-			// Standard Error: 26_835
-			.saturating_add(Weight::from_parts(157_003, 0).saturating_mul(s.into()))
+		// Minimum execution time: 89_674_000 picoseconds.
+		Weight::from_parts(92_195_302, 6531)
+			// Standard Error: 13_523
+			.saturating_add(Weight::from_parts(121_658, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(8_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -384,8 +384,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `161`
 		//  Estimated: `4177`
-		// Minimum execution time: 13_962_000 picoseconds.
-		Weight::from_parts(14_566_000, 4177)
+		// Minimum execution time: 13_045_000 picoseconds.
+		Weight::from_parts(13_415_000, 4177)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -397,8 +397,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `67`
 		//  Estimated: `4008`
-		// Minimum execution time: 10_804_000 picoseconds.
-		Weight::from_parts(11_092_000, 4008)
+		// Minimum execution time: 10_533_000 picoseconds.
+		Weight::from_parts(10_914_000, 4008)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -410,8 +410,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `67`
 		//  Estimated: `4008`
-		// Minimum execution time: 10_857_000 picoseconds.
-		Weight::from_parts(11_268_000, 4008)
+		// Minimum execution time: 10_301_000 picoseconds.
+		Weight::from_parts(10_646_000, 4008)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -429,8 +429,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `147`
 		//  Estimated: `4107`
-		// Minimum execution time: 19_016_000 picoseconds.
-		Weight::from_parts(19_618_000, 4107)
+		// Minimum execution time: 18_100_000 picoseconds.
+		Weight::from_parts(18_367_000, 4107)
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -445,10 +445,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `316 + s * (6 ±0)`
 		//  Estimated: `4177`
-		// Minimum execution time: 16_854_000 picoseconds.
-		Weight::from_parts(17_824_532, 4177)
-			// Standard Error: 3_956
-			.saturating_add(Weight::from_parts(111_331, 0).saturating_mul(s.into()))
+		// Minimum execution time: 16_195_000 picoseconds.
+		Weight::from_parts(17_118_368, 4177)
+			// Standard Error: 4_262
+			.saturating_add(Weight::from_parts(108_183, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -461,6 +461,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
