@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_schemas`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -61,10 +61,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 12_307_000 picoseconds.
-		Weight::from_parts(12_715_000, 1984)
-			// Standard Error: 34
-			.saturating_add(Weight::from_parts(34_197, 0).saturating_mul(m.into()))
+		// Minimum execution time: 11_836_000 picoseconds.
+		Weight::from_parts(12_010_000, 1984)
+			// Standard Error: 48
+			.saturating_add(Weight::from_parts(34_625, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -81,10 +81,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 12_277_000 picoseconds.
-		Weight::from_parts(12_505_000, 1984)
-			// Standard Error: 35
-			.saturating_add(Weight::from_parts(34_255, 0).saturating_mul(m.into()))
+		// Minimum execution time: 11_944_000 picoseconds.
+		Weight::from_parts(12_202_000, 1984)
+			// Standard Error: 48
+			.saturating_add(Weight::from_parts(34_837, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -103,10 +103,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `230`
 		//  Estimated: `4190`
-		// Minimum execution time: 16_628_000 picoseconds.
-		Weight::from_parts(4_241_944, 4190)
-			// Standard Error: 35
-			.saturating_add(Weight::from_parts(3_182, 0).saturating_mul(m.into()))
+		// Minimum execution time: 16_869_000 picoseconds.
+		Weight::from_parts(4_420_065, 4190)
+			// Standard Error: 34
+			.saturating_add(Weight::from_parts(3_176, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -123,10 +123,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 13_111_000 picoseconds.
-		Weight::from_parts(13_336_000, 1984)
-			// Standard Error: 34
-			.saturating_add(Weight::from_parts(34_082, 0).saturating_mul(m.into()))
+		// Minimum execution time: 12_032_000 picoseconds.
+		Weight::from_parts(12_229_000, 1984)
+			// Standard Error: 47
+			.saturating_add(Weight::from_parts(34_728, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -145,10 +145,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `4562`
-		// Minimum execution time: 20_273_000 picoseconds.
-		Weight::from_parts(5_715_450, 4562)
-			// Standard Error: 56
-			.saturating_add(Weight::from_parts(34_505, 0).saturating_mul(m.into()))
+		// Minimum execution time: 19_869_000 picoseconds.
+		Weight::from_parts(19_995_000, 4562)
+			// Standard Error: 45
+			.saturating_add(Weight::from_parts(34_820, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -158,8 +158,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 4_118_000 picoseconds.
-		Weight::from_parts(4_347_000, 0)
+		// Minimum execution time: 4_117_000 picoseconds.
+		Weight::from_parts(4_346_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	/// Storage: `Schemas::GovernanceSchemaModelMaxBytes` (r:1 w:0)
@@ -177,10 +177,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `4562`
-		// Minimum execution time: 19_878_000 picoseconds.
-		Weight::from_parts(4_574_505, 4562)
-			// Standard Error: 56
-			.saturating_add(Weight::from_parts(34_237, 0).saturating_mul(m.into()))
+		// Minimum execution time: 19_918_000 picoseconds.
+		Weight::from_parts(20_091_000, 4562)
+			// Standard Error: 50
+			.saturating_add(Weight::from_parts(34_210, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -199,10 +199,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `230`
 		//  Estimated: `4190`
-		// Minimum execution time: 16_945_000 picoseconds.
-		Weight::from_parts(5_313_648, 4190)
-			// Standard Error: 33
-			.saturating_add(Weight::from_parts(3_151, 0).saturating_mul(m.into()))
+		// Minimum execution time: 17_116_000 picoseconds.
+		Weight::from_parts(5_248_919, 4190)
+			// Standard Error: 32
+			.saturating_add(Weight::from_parts(3_155, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -222,8 +222,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `433`
 		//  Estimated: `4393`
-		// Minimum execution time: 22_765_000 picoseconds.
-		Weight::from_parts(23_075_000, 4393)
+		// Minimum execution time: 22_403_000 picoseconds.
+		Weight::from_parts(23_052_000, 4393)
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -235,8 +235,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `203`
 		//  Estimated: `4562`
-		// Minimum execution time: 14_593_000 picoseconds.
-		Weight::from_parts(15_172_000, 4562)
+		// Minimum execution time: 14_578_000 picoseconds.
+		Weight::from_parts(15_023_000, 4562)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -257,10 +257,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 12_307_000 picoseconds.
-		Weight::from_parts(12_715_000, 1984)
-			// Standard Error: 34
-			.saturating_add(Weight::from_parts(34_197, 0).saturating_mul(m.into()))
+		// Minimum execution time: 11_836_000 picoseconds.
+		Weight::from_parts(12_010_000, 1984)
+			// Standard Error: 48
+			.saturating_add(Weight::from_parts(34_625, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -277,10 +277,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 12_277_000 picoseconds.
-		Weight::from_parts(12_505_000, 1984)
-			// Standard Error: 35
-			.saturating_add(Weight::from_parts(34_255, 0).saturating_mul(m.into()))
+		// Minimum execution time: 11_944_000 picoseconds.
+		Weight::from_parts(12_202_000, 1984)
+			// Standard Error: 48
+			.saturating_add(Weight::from_parts(34_837, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -299,10 +299,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `230`
 		//  Estimated: `4190`
-		// Minimum execution time: 16_628_000 picoseconds.
-		Weight::from_parts(4_241_944, 4190)
-			// Standard Error: 35
-			.saturating_add(Weight::from_parts(3_182, 0).saturating_mul(m.into()))
+		// Minimum execution time: 16_869_000 picoseconds.
+		Weight::from_parts(4_420_065, 4190)
+			// Standard Error: 34
+			.saturating_add(Weight::from_parts(3_176, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -319,10 +319,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `1984`
-		// Minimum execution time: 13_111_000 picoseconds.
-		Weight::from_parts(13_336_000, 1984)
-			// Standard Error: 34
-			.saturating_add(Weight::from_parts(34_082, 0).saturating_mul(m.into()))
+		// Minimum execution time: 12_032_000 picoseconds.
+		Weight::from_parts(12_229_000, 1984)
+			// Standard Error: 47
+			.saturating_add(Weight::from_parts(34_728, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -341,10 +341,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `4562`
-		// Minimum execution time: 20_273_000 picoseconds.
-		Weight::from_parts(5_715_450, 4562)
-			// Standard Error: 56
-			.saturating_add(Weight::from_parts(34_505, 0).saturating_mul(m.into()))
+		// Minimum execution time: 19_869_000 picoseconds.
+		Weight::from_parts(19_995_000, 4562)
+			// Standard Error: 45
+			.saturating_add(Weight::from_parts(34_820, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -354,8 +354,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 4_118_000 picoseconds.
-		Weight::from_parts(4_347_000, 0)
+		// Minimum execution time: 4_117_000 picoseconds.
+		Weight::from_parts(4_346_000, 0)
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	/// Storage: `Schemas::GovernanceSchemaModelMaxBytes` (r:1 w:0)
@@ -373,10 +373,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `136`
 		//  Estimated: `4562`
-		// Minimum execution time: 19_878_000 picoseconds.
-		Weight::from_parts(4_574_505, 4562)
-			// Standard Error: 56
-			.saturating_add(Weight::from_parts(34_237, 0).saturating_mul(m.into()))
+		// Minimum execution time: 19_918_000 picoseconds.
+		Weight::from_parts(20_091_000, 4562)
+			// Standard Error: 50
+			.saturating_add(Weight::from_parts(34_210, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -395,10 +395,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `230`
 		//  Estimated: `4190`
-		// Minimum execution time: 16_945_000 picoseconds.
-		Weight::from_parts(5_313_648, 4190)
-			// Standard Error: 33
-			.saturating_add(Weight::from_parts(3_151, 0).saturating_mul(m.into()))
+		// Minimum execution time: 17_116_000 picoseconds.
+		Weight::from_parts(5_248_919, 4190)
+			// Standard Error: 32
+			.saturating_add(Weight::from_parts(3_155, 0).saturating_mul(m.into()))
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -418,8 +418,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `433`
 		//  Estimated: `4393`
-		// Minimum execution time: 22_765_000 picoseconds.
-		Weight::from_parts(23_075_000, 4393)
+		// Minimum execution time: 22_403_000 picoseconds.
+		Weight::from_parts(23_052_000, 4393)
 			.saturating_add(RocksDbWeight::get().reads(5_u64))
 			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
@@ -431,8 +431,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `203`
 		//  Estimated: `4562`
-		// Minimum execution time: 14_593_000 picoseconds.
-		Weight::from_parts(15_172_000, 4562)
+		// Minimum execution time: 14_578_000 picoseconds.
+		Weight::from_parts(15_023_000, 4562)
 			.saturating_add(RocksDbWeight::get().reads(2_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -445,6 +445,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
