@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_stateful_storage`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -59,10 +59,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `318`
 		//  Estimated: `5763`
-		// Minimum execution time: 25_521_000 picoseconds.
-		Weight::from_parts(25_009_834, 5763)
-			// Standard Error: 50
-			.saturating_add(Weight::from_parts(1_095, 0).saturating_mul(s.into()))
+		// Minimum execution time: 25_067_000 picoseconds.
+		Weight::from_parts(24_361_305, 5763)
+			// Standard Error: 45
+			.saturating_add(Weight::from_parts(989, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -79,8 +79,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `10696`
 		//  Estimated: `16141`
-		// Minimum execution time: 39_580_000 picoseconds.
-		Weight::from_parts(43_579_447, 16141)
+		// Minimum execution time: 39_119_000 picoseconds.
+		Weight::from_parts(43_026_621, 16141)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -97,10 +97,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `1479`
 		//  Estimated: `6924`
-		// Minimum execution time: 29_800_000 picoseconds.
-		Weight::from_parts(30_868_152, 6924)
-			// Standard Error: 153
-			.saturating_add(Weight::from_parts(639, 0).saturating_mul(s.into()))
+		// Minimum execution time: 29_371_000 picoseconds.
+		Weight::from_parts(30_696_652, 6924)
+			// Standard Error: 244
+			.saturating_add(Weight::from_parts(607, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -116,8 +116,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `1477`
 		//  Estimated: `6922`
-		// Minimum execution time: 28_360_000 picoseconds.
-		Weight::from_parts(29_206_000, 6922)
+		// Minimum execution time: 28_199_000 picoseconds.
+		Weight::from_parts(28_924_000, 6922)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -132,10 +132,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `325`
 		//  Estimated: `5770`
-		// Minimum execution time: 71_848_000 picoseconds.
-		Weight::from_parts(64_366_301, 5770)
-			// Standard Error: 180
-			.saturating_add(Weight::from_parts(6_289, 0).saturating_mul(s.into()))
+		// Minimum execution time: 71_641_000 picoseconds.
+		Weight::from_parts(64_412_830, 5770)
+			// Standard Error: 181
+			.saturating_add(Weight::from_parts(6_261, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -150,8 +150,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `10703`
 		//  Estimated: `16148`
-		// Minimum execution time: 79_847_000 picoseconds.
-		Weight::from_parts(85_226_998, 16148)
+		// Minimum execution time: 80_538_000 picoseconds.
+		Weight::from_parts(84_383_821, 16148)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -166,10 +166,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `1412`
 		//  Estimated: `6857`
-		// Minimum execution time: 69_282_000 picoseconds.
-		Weight::from_parts(70_072_021, 6857)
-			// Standard Error: 291
-			.saturating_add(Weight::from_parts(6_551, 0).saturating_mul(s.into()))
+		// Minimum execution time: 68_855_000 picoseconds.
+		Weight::from_parts(70_897_498, 6857)
+			// Standard Error: 271
+			.saturating_add(Weight::from_parts(5_998, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -183,8 +183,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `1410`
 		//  Estimated: `6855`
-		// Minimum execution time: 67_486_000 picoseconds.
-		Weight::from_parts(68_873_000, 6855)
+		// Minimum execution time: 67_709_000 picoseconds.
+		Weight::from_parts(69_306_000, 6855)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -205,10 +205,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `318`
 		//  Estimated: `5763`
-		// Minimum execution time: 25_521_000 picoseconds.
-		Weight::from_parts(25_009_834, 5763)
-			// Standard Error: 50
-			.saturating_add(Weight::from_parts(1_095, 0).saturating_mul(s.into()))
+		// Minimum execution time: 25_067_000 picoseconds.
+		Weight::from_parts(24_361_305, 5763)
+			// Standard Error: 45
+			.saturating_add(Weight::from_parts(989, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -225,8 +225,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `10696`
 		//  Estimated: `16141`
-		// Minimum execution time: 39_580_000 picoseconds.
-		Weight::from_parts(43_579_447, 16141)
+		// Minimum execution time: 39_119_000 picoseconds.
+		Weight::from_parts(43_026_621, 16141)
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -243,10 +243,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `1479`
 		//  Estimated: `6924`
-		// Minimum execution time: 29_800_000 picoseconds.
-		Weight::from_parts(30_868_152, 6924)
-			// Standard Error: 153
-			.saturating_add(Weight::from_parts(639, 0).saturating_mul(s.into()))
+		// Minimum execution time: 29_371_000 picoseconds.
+		Weight::from_parts(30_696_652, 6924)
+			// Standard Error: 244
+			.saturating_add(Weight::from_parts(607, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -262,8 +262,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `1477`
 		//  Estimated: `6922`
-		// Minimum execution time: 28_360_000 picoseconds.
-		Weight::from_parts(29_206_000, 6922)
+		// Minimum execution time: 28_199_000 picoseconds.
+		Weight::from_parts(28_924_000, 6922)
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -278,10 +278,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `325`
 		//  Estimated: `5770`
-		// Minimum execution time: 71_848_000 picoseconds.
-		Weight::from_parts(64_366_301, 5770)
-			// Standard Error: 180
-			.saturating_add(Weight::from_parts(6_289, 0).saturating_mul(s.into()))
+		// Minimum execution time: 71_641_000 picoseconds.
+		Weight::from_parts(64_412_830, 5770)
+			// Standard Error: 181
+			.saturating_add(Weight::from_parts(6_261, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -296,8 +296,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `10703`
 		//  Estimated: `16148`
-		// Minimum execution time: 79_847_000 picoseconds.
-		Weight::from_parts(85_226_998, 16148)
+		// Minimum execution time: 80_538_000 picoseconds.
+		Weight::from_parts(84_383_821, 16148)
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -312,10 +312,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `1412`
 		//  Estimated: `6857`
-		// Minimum execution time: 69_282_000 picoseconds.
-		Weight::from_parts(70_072_021, 6857)
-			// Standard Error: 291
-			.saturating_add(Weight::from_parts(6_551, 0).saturating_mul(s.into()))
+		// Minimum execution time: 68_855_000 picoseconds.
+		Weight::from_parts(70_897_498, 6857)
+			// Standard Error: 271
+			.saturating_add(Weight::from_parts(5_998, 0).saturating_mul(s.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -329,8 +329,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `1410`
 		//  Estimated: `6855`
-		// Minimum execution time: 67_486_000 picoseconds.
-		Weight::from_parts(68_873_000, 6855)
+		// Minimum execution time: 67_709_000 picoseconds.
+		Weight::from_parts(69_306_000, 6855)
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
@@ -343,6 +343,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
