@@ -2,9 +2,9 @@
 //! Autogenerated weights for `pallet_time_release`
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `20`, REPEAT: `10`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! WASM-EXECUTION: `Compiled`, CHAIN: `Some("frequency-bench")`, DB CACHE: `1024`
 
 // Executed Command:
@@ -57,8 +57,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `113`
 		//  Estimated: `5409`
-		// Minimum execution time: 42_382_000 picoseconds.
-		Weight::from_parts(42_932_000, 5409)
+		// Minimum execution time: 41_626_000 picoseconds.
+		Weight::from_parts(42_667_000, 5409)
 			.saturating_add(T::DbWeight::get().reads(6_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -75,10 +75,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `268`
 		//  Estimated: `5409`
-		// Minimum execution time: 27_483_000 picoseconds.
-		Weight::from_parts(28_196_769, 5409)
-			// Standard Error: 1_910
-			.saturating_add(Weight::from_parts(1_679, 0).saturating_mul(i.into()))
+		// Minimum execution time: 26_897_000 picoseconds.
+		Weight::from_parts(27_668_174, 5409)
+			// Standard Error: 2_213
+			.saturating_add(Weight::from_parts(9_813, 0).saturating_mul(i.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -95,10 +95,10 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `39`
 		//  Estimated: `5259`
-		// Minimum execution time: 21_782_000 picoseconds.
-		Weight::from_parts(22_466_489, 5259)
-			// Standard Error: 1_849
-			.saturating_add(Weight::from_parts(57_259, 0).saturating_mul(i.into()))
+		// Minimum execution time: 21_220_000 picoseconds.
+		Weight::from_parts(22_365_488, 5259)
+			// Standard Error: 1_975
+			.saturating_add(Weight::from_parts(57_298, 0).saturating_mul(i.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -122,8 +122,8 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `113`
 		//  Estimated: `5409`
-		// Minimum execution time: 42_382_000 picoseconds.
-		Weight::from_parts(42_932_000, 5409)
+		// Minimum execution time: 41_626_000 picoseconds.
+		Weight::from_parts(42_667_000, 5409)
 			.saturating_add(RocksDbWeight::get().reads(6_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -140,10 +140,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `268`
 		//  Estimated: `5409`
-		// Minimum execution time: 27_483_000 picoseconds.
-		Weight::from_parts(28_196_769, 5409)
-			// Standard Error: 1_910
-			.saturating_add(Weight::from_parts(1_679, 0).saturating_mul(i.into()))
+		// Minimum execution time: 26_897_000 picoseconds.
+		Weight::from_parts(27_668_174, 5409)
+			// Standard Error: 2_213
+			.saturating_add(Weight::from_parts(9_813, 0).saturating_mul(i.into()))
 			.saturating_add(RocksDbWeight::get().reads(4_u64))
 			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
@@ -160,10 +160,10 @@ impl WeightInfo for () {
 		// Proof Size summary in bytes:
 		//  Measured:  `39`
 		//  Estimated: `5259`
-		// Minimum execution time: 21_782_000 picoseconds.
-		Weight::from_parts(22_466_489, 5259)
-			// Standard Error: 1_849
-			.saturating_add(Weight::from_parts(57_259, 0).saturating_mul(i.into()))
+		// Minimum execution time: 21_220_000 picoseconds.
+		Weight::from_parts(22_365_488, 5259)
+			// Standard Error: 1_975
+			.saturating_add(Weight::from_parts(57_298, 0).saturating_mul(i.into()))
 			.saturating_add(RocksDbWeight::get().reads(3_u64))
 			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
@@ -176,6 +176,7 @@ mod tests {
   use common_runtime::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use common_runtime::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
+  #[allow(dead_code)]
   struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
