@@ -24,7 +24,7 @@ pub type MaxSchemaGrants = ConstU32<30>;
 /// up by `pallet_aura` to implement `fn slot_duration()`.
 ///
 /// Change this to adjust the block time.
-pub const MILLISECS_PER_BLOCK: u64 = 12000;
+pub const MILLISECS_PER_BLOCK: u64 = prod_or_testnet_or_local!(12_000, 6_000, 6_000);
 
 // NOTE: Currently it is not possible to change the slot duration after the chain has started.
 //       Attempting to do so will brick block production.
@@ -63,10 +63,19 @@ pub const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_percent(5);
 /// `Operational` extrinsics.
 pub const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
 
+#[cfg(not(any(feature = "frequency-testnet", feature = "frequency-local")))]
 /// We allow for 0.5 of a second of compute with a 12 second average block time.
 pub const MAXIMUM_BLOCK_WEIGHT: Weight = Weight::from_parts(WEIGHT_REF_TIME_PER_SECOND, 0)
 	.saturating_div(2)
 	.set_proof_size(cumulus_primitives_core::relay_chain::MAX_POV_SIZE as u64);
+
+#[cfg(any(feature = "frequency-testnet", feature = "frequency-local"))]
+/// We allow for 2 seconds of compute with a 6 second average block time.
+pub const MAXIMUM_BLOCK_WEIGHT: Weight = Weight::from_parts(
+	WEIGHT_REF_TIME_PER_SECOND.saturating_mul(2),
+	cumulus_primitives_core::relay_chain::MAX_POV_SIZE as u64,
+);
+
 pub type ZERO = ConstU32<0>;
 pub type TWO = ConstU32<2>;
 pub type FIFTY = ConstU32<50>;
@@ -125,6 +134,11 @@ pub type MinReleaseTransfer = ConstU128<0>;
 pub const MAX_RELEASE_SCHEDULES: u32 = 50;
 // -end- TimeRelease Pallet ---
 
+#[cfg(any(feature = "frequency-testnet", feature = "frequency-local"))]
+// --- Timestamp Pallet ---
+pub type MinimumPeriod = ConstU64<0>;
+
+#[cfg(not(any(feature = "frequency-testnet", feature = "frequency-local")))]
 // --- Timestamp Pallet ---
 pub type MinimumPeriod = ConstU64<{ SLOT_DURATION / 2 }>;
 // -end- Timestamp Pallet ---
