@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_balances
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -39,8 +39,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `4088`
-		// Minimum execution time: 44_286_000 picoseconds.
-		Weight::from_parts(45_164_000, 4088)
+		// Minimum execution time: 43_078_000 picoseconds.
+		Weight::from_parts(43_628_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -50,8 +50,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `4088`
-		// Minimum execution time: 35_627_000 picoseconds.
-		Weight::from_parts(36_295_000, 4088)
+		// Minimum execution time: 34_392_000 picoseconds.
+		Weight::from_parts(34_985_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -61,8 +61,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `207`
 		//  Estimated: `4088`
-		// Minimum execution time: 13_030_000 picoseconds.
-		Weight::from_parts(13_432_000, 4088)
+		// Minimum execution time: 12_662_000 picoseconds.
+		Weight::from_parts(13_067_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -72,8 +72,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `207`
 		//  Estimated: `4088`
-		// Minimum execution time: 17_974_000 picoseconds.
-		Weight::from_parts(18_546_000, 4088)
+		// Minimum execution time: 17_799_000 picoseconds.
+		Weight::from_parts(18_316_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -83,8 +83,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `103`
 		//  Estimated: `6691`
-		// Minimum execution time: 46_663_000 picoseconds.
-		Weight::from_parts(47_281_000, 6691)
+		// Minimum execution time: 45_102_000 picoseconds.
+		Weight::from_parts(45_888_000, 6691)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -94,8 +94,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `4088`
-		// Minimum execution time: 44_283_000 picoseconds.
-		Weight::from_parts(45_189_000, 4088)
+		// Minimum execution time: 42_953_000 picoseconds.
+		Weight::from_parts(43_391_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -105,8 +105,8 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `207`
 		//  Estimated: `4088`
-		// Minimum execution time: 16_071_000 picoseconds.
-		Weight::from_parts(16_458_000, 4088)
+		// Minimum execution time: 15_522_000 picoseconds.
+		Weight::from_parts(16_045_000, 4088)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -117,10 +117,10 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0 + u * (136 ±0)`
 		//  Estimated: `1485 + u * (2603 ±0)`
-		// Minimum execution time: 15_547_000 picoseconds.
-		Weight::from_parts(15_791_000, 1485)
-			// Standard Error: 12_930
-			.saturating_add(Weight::from_parts(13_735_185, 0).saturating_mul(u.into()))
+		// Minimum execution time: 14_875_000 picoseconds.
+		Weight::from_parts(15_232_000, 1485)
+			// Standard Error: 10_643
+			.saturating_add(Weight::from_parts(13_143_686, 0).saturating_mul(u.into()))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(u.into())))
 			.saturating_add(T::DbWeight::get().writes((1_u64).saturating_mul(u.into())))
 			.saturating_add(Weight::from_parts(0, 2603).saturating_mul(u.into()))
@@ -129,8 +129,22 @@ impl<T: frame_system::Config> pallet_balances::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 4_955_000 picoseconds.
-		Weight::from_parts(5_331_000, 0)
+		// Minimum execution time: 4_840_000 picoseconds.
+		Weight::from_parts(5_052_000, 0)
+	}
+	fn burn_allow_death() -> Weight {
+		// Proof Size summary in bytes:
+		//  Measured:  `0`
+		//  Estimated: `0`
+		// Minimum execution time: 26_719_000 picoseconds.
+		Weight::from_parts(27_053_000, 0)
+	}
+	fn burn_keep_alive() -> Weight {
+		// Proof Size summary in bytes:
+		//  Measured:  `0`
+		//  Estimated: `0`
+		// Minimum execution time: 18_009_000 picoseconds.
+		Weight::from_parts(18_469_000, 0)
 	}
 }
 
