@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_collator_selection
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -42,10 +42,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `334 + b * (79 ±0)`
 		//  Estimated: `1819 + b * (2554 ±0)`
-		// Minimum execution time: 10_758_000 picoseconds.
-		Weight::from_parts(9_214_948, 1819)
-			// Standard Error: 9_360
-			.saturating_add(Weight::from_parts(3_174_494, 0).saturating_mul(b.into()))
+		// Minimum execution time: 10_654_000 picoseconds.
+		Weight::from_parts(9_095_008, 1819)
+			// Standard Error: 9_095
+			.saturating_add(Weight::from_parts(3_145_323, 0).saturating_mul(b.into()))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(b.into())))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 			.saturating_add(Weight::from_parts(0, 2554).saturating_mul(b.into()))
@@ -64,10 +64,12 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `1120 + b * (32 ±0) + c * (51 ±0)`
 		//  Estimated: `5025 + b * (29 ±0) + c * (53 ±0)`
-		// Minimum execution time: 40_521_000 picoseconds.
-		Weight::from_parts(41_841_860, 5025)
-			// Standard Error: 1_464
-			.saturating_add(Weight::from_parts(183_911, 0).saturating_mul(c.into()))
+		// Minimum execution time: 40_869_000 picoseconds.
+		Weight::from_parts(41_654_262, 5025)
+			// Standard Error: 4_382
+			.saturating_add(Weight::from_parts(8_221, 0).saturating_mul(b.into()))
+			// Standard Error: 1_299
+			.saturating_add(Weight::from_parts(163_499, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 29).saturating_mul(b.into()))
@@ -82,10 +84,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `153 + b * (32 ±0)`
 		//  Estimated: `4381`
-		// Minimum execution time: 10_128_000 picoseconds.
-		Weight::from_parts(10_273_003, 4381)
-			// Standard Error: 1_944
-			.saturating_add(Weight::from_parts(162_607, 0).saturating_mul(b.into()))
+		// Minimum execution time: 9_958_000 picoseconds.
+		Weight::from_parts(10_224_998, 4381)
+			// Standard Error: 2_183
+			.saturating_add(Weight::from_parts(163_598, 0).saturating_mul(b.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -95,8 +97,8 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_894_000 picoseconds.
-		Weight::from_parts(4_326_000, 0)
+		// Minimum execution time: 3_729_000 picoseconds.
+		Weight::from_parts(4_180_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	/// Storage: `CollatorSelection::CandidacyBond` (r:1 w:1)
@@ -113,12 +115,12 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `0 + c * (188 ±0) + k * (130 ±0)`
 		//  Estimated: `4381 + c * (894 ±29) + k * (894 ±29)`
-		// Minimum execution time: 8_480_000 picoseconds.
-		Weight::from_parts(8_741_000, 4381)
-			// Standard Error: 152_640
-			.saturating_add(Weight::from_parts(5_301_722, 0).saturating_mul(c.into()))
-			// Standard Error: 152_640
-			.saturating_add(Weight::from_parts(4_976_098, 0).saturating_mul(k.into()))
+		// Minimum execution time: 8_318_000 picoseconds.
+		Weight::from_parts(8_452_000, 4381)
+			// Standard Error: 153_809
+			.saturating_add(Weight::from_parts(5_311_760, 0).saturating_mul(c.into()))
+			// Standard Error: 153_809
+			.saturating_add(Weight::from_parts(5_020_659, 0).saturating_mul(k.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 			.saturating_add(T::DbWeight::get().writes((1_u64).saturating_mul(c.into())))
@@ -135,10 +137,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `372 + c * (49 ±0)`
 		//  Estimated: `4381`
-		// Minimum execution time: 21_997_000 picoseconds.
-		Weight::from_parts(24_301_403, 4381)
-			// Standard Error: 2_181
-			.saturating_add(Weight::from_parts(189_193, 0).saturating_mul(c.into()))
+		// Minimum execution time: 23_007_000 picoseconds.
+		Weight::from_parts(25_300_538, 4381)
+			// Standard Error: 2_173
+			.saturating_add(Weight::from_parts(178_620, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -157,10 +159,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `810 + c * (57 ±0)`
 		//  Estimated: `4774 + c * (57 ±0)`
-		// Minimum execution time: 33_496_000 picoseconds.
-		Weight::from_parts(36_040_937, 4774)
-			// Standard Error: 3_236
-			.saturating_add(Weight::from_parts(257_787, 0).saturating_mul(c.into()))
+		// Minimum execution time: 35_180_000 picoseconds.
+		Weight::from_parts(37_653_329, 4774)
+			// Standard Error: 3_057
+			.saturating_add(Weight::from_parts(232_265, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 			.saturating_add(Weight::from_parts(0, 57).saturating_mul(c.into()))
@@ -182,10 +184,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `901 + c * (58 ±0)`
 		//  Estimated: `4866 + c * (59 ±0)`
-		// Minimum execution time: 49_067_000 picoseconds.
-		Weight::from_parts(51_631_074, 4866)
-			// Standard Error: 3_777
-			.saturating_add(Weight::from_parts(278_968, 0).saturating_mul(c.into()))
+		// Minimum execution time: 49_376_000 picoseconds.
+		Weight::from_parts(53_285_478, 4866)
+			// Standard Error: 3_622
+			.saturating_add(Weight::from_parts(258_728, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 			.saturating_add(Weight::from_parts(0, 59).saturating_mul(c.into()))
@@ -201,10 +203,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `501 + c * (48 ±0)`
 		//  Estimated: `4381`
-		// Minimum execution time: 25_988_000 picoseconds.
-		Weight::from_parts(28_188_871, 4381)
-			// Standard Error: 2_899
-			.saturating_add(Weight::from_parts(228_280, 0).saturating_mul(c.into()))
+		// Minimum execution time: 26_219_000 picoseconds.
+		Weight::from_parts(28_384_573, 4381)
+			// Standard Error: 2_502
+			.saturating_add(Weight::from_parts(207_343, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -216,8 +218,8 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `212`
 		//  Estimated: `6691`
-		// Minimum execution time: 38_632_000 picoseconds.
-		Weight::from_parts(39_725_000, 6691)
+		// Minimum execution time: 39_059_000 picoseconds.
+		Weight::from_parts(39_658_000, 6691)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -237,10 +239,10 @@ impl<T: frame_system::Config> pallet_collator_selection::WeightInfo for Substrat
 		// Proof Size summary in bytes:
 		//  Measured:  `1188 + c * (98 ±0) + r * (128 ±0)`
 		//  Estimated: `4381 + c * (2519 ±0) + r * (2603 ±0)`
-		// Minimum execution time: 17_741_000 picoseconds.
-		Weight::from_parts(18_072_000, 4381)
-			// Standard Error: 277_992
-			.saturating_add(Weight::from_parts(12_316_084, 0).saturating_mul(c.into()))
+		// Minimum execution time: 17_445_000 picoseconds.
+		Weight::from_parts(17_685_000, 4381)
+			// Standard Error: 273_806
+			.saturating_add(Weight::from_parts(12_442_922, 0).saturating_mul(c.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(c.into())))
 			.saturating_add(T::DbWeight::get().writes((1_u64).saturating_mul(c.into())))
