@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_collective
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -51,12 +51,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `0 + m * (832 ±0) + p * (310 ±0)`
 		//  Estimated: `4789 + m * (489 ±3) + p * (2615 ±1)`
-		// Minimum execution time: 7_328_000 picoseconds.
-		Weight::from_parts(7_573_000, 4789)
-			// Standard Error: 63_163
-			.saturating_add(Weight::from_parts(2_961_099, 0).saturating_mul(m.into()))
-			// Standard Error: 25_667
-			.saturating_add(Weight::from_parts(3_397_864, 0).saturating_mul(p.into()))
+		// Minimum execution time: 7_042_000 picoseconds.
+		Weight::from_parts(7_344_000, 4789)
+			// Standard Error: 62_217
+			.saturating_add(Weight::from_parts(2_954_755, 0).saturating_mul(m.into()))
+			// Standard Error: 25_283
+			.saturating_add(Weight::from_parts(3_340_034, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(p.into())))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
@@ -74,12 +74,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `70 + m * (32 ±0)`
 		//  Estimated: `2049 + m * (32 ±0)`
-		// Minimum execution time: 11_041_000 picoseconds.
-		Weight::from_parts(11_285_463, 2049)
-			// Standard Error: 14
-			.saturating_add(Weight::from_parts(1_519, 0).saturating_mul(b.into()))
-			// Standard Error: 1_577
-			.saturating_add(Weight::from_parts(26_044, 0).saturating_mul(m.into()))
+		// Minimum execution time: 10_753_000 picoseconds.
+		Weight::from_parts(10_928_530, 2049)
+			// Standard Error: 11
+			.saturating_add(Weight::from_parts(1_517, 0).saturating_mul(b.into()))
+			// Standard Error: 1_266
+			.saturating_add(Weight::from_parts(28_654, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(m.into()))
 	}
@@ -95,12 +95,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `70 + m * (32 ±0)`
 		//  Estimated: `4029 + m * (32 ±0)`
-		// Minimum execution time: 13_221_000 picoseconds.
-		Weight::from_parts(13_619_342, 4029)
-			// Standard Error: 16
-			.saturating_add(Weight::from_parts(1_554, 0).saturating_mul(b.into()))
-			// Standard Error: 1_710
-			.saturating_add(Weight::from_parts(34_756, 0).saturating_mul(m.into()))
+		// Minimum execution time: 13_039_000 picoseconds.
+		Weight::from_parts(13_171_547, 4029)
+			// Standard Error: 15
+			.saturating_add(Weight::from_parts(1_614, 0).saturating_mul(b.into()))
+			// Standard Error: 1_665
+			.saturating_add(Weight::from_parts(43_747, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(m.into()))
 	}
@@ -124,14 +124,14 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `113 + m * (32 ±0) + p * (51 ±0)`
 		//  Estimated: `4025 + m * (42 ±0) + p * (49 ±0)`
-		// Minimum execution time: 18_154_000 picoseconds.
-		Weight::from_parts(17_073_803, 4025)
-			// Standard Error: 44
-			.saturating_add(Weight::from_parts(2_330, 0).saturating_mul(b.into()))
-			// Standard Error: 5_271
-			.saturating_add(Weight::from_parts(147_322, 0).saturating_mul(m.into()))
-			// Standard Error: 1_839
-			.saturating_add(Weight::from_parts(252_960, 0).saturating_mul(p.into()))
+		// Minimum execution time: 17_883_000 picoseconds.
+		Weight::from_parts(16_770_176, 4025)
+			// Standard Error: 45
+			.saturating_add(Weight::from_parts(2_202, 0).saturating_mul(b.into()))
+			// Standard Error: 5_427
+			.saturating_add(Weight::from_parts(157_016, 0).saturating_mul(m.into()))
+			// Standard Error: 1_894
+			.saturating_add(Weight::from_parts(267_891, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 			.saturating_add(Weight::from_parts(0, 42).saturating_mul(m.into()))
@@ -147,10 +147,10 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `571 + m * (64 ±0)`
 		//  Estimated: `4531 + m * (64 ±0)`
-		// Minimum execution time: 16_248_000 picoseconds.
-		Weight::from_parts(17_022_642, 4531)
-			// Standard Error: 4_789
-			.saturating_add(Weight::from_parts(67_021, 0).saturating_mul(m.into()))
+		// Minimum execution time: 15_902_000 picoseconds.
+		Weight::from_parts(16_382_607, 4531)
+			// Standard Error: 5_751
+			.saturating_add(Weight::from_parts(113_517, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 			.saturating_add(Weight::from_parts(0, 64).saturating_mul(m.into()))
@@ -171,12 +171,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `41 + m * (64 ±0) + p * (54 ±0)`
 		//  Estimated: `4019 + m * (80 ±0) + p * (50 ±0)`
-		// Minimum execution time: 18_755_000 picoseconds.
-		Weight::from_parts(18_868_734, 4019)
-			// Standard Error: 5_883
-			.saturating_add(Weight::from_parts(139_895, 0).saturating_mul(m.into()))
-			// Standard Error: 1_569
-			.saturating_add(Weight::from_parts(223_209, 0).saturating_mul(p.into()))
+		// Minimum execution time: 18_412_000 picoseconds.
+		Weight::from_parts(18_635_993, 4019)
+			// Standard Error: 5_477
+			.saturating_add(Weight::from_parts(151_104, 0).saturating_mul(m.into()))
+			// Standard Error: 1_461
+			.saturating_add(Weight::from_parts(241_387, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 80).saturating_mul(m.into()))
@@ -200,14 +200,14 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `72 + b * (1 ±0) + m * (64 ±0) + p * (72 ±0)`
 		//  Estimated: `4025 + b * (1 ±0) + m * (85 ±0) + p * (65 ±0)`
-		// Minimum execution time: 27_825_000 picoseconds.
-		Weight::from_parts(28_624_476, 4025)
-			// Standard Error: 64
-			.saturating_add(Weight::from_parts(1_817, 0).saturating_mul(b.into()))
-			// Standard Error: 9_935
-			.saturating_add(Weight::from_parts(96_088, 0).saturating_mul(m.into()))
-			// Standard Error: 2_658
-			.saturating_add(Weight::from_parts(389_593, 0).saturating_mul(p.into()))
+		// Minimum execution time: 28_153_000 picoseconds.
+		Weight::from_parts(28_081_508, 4025)
+			// Standard Error: 61
+			.saturating_add(Weight::from_parts(1_901, 0).saturating_mul(b.into()))
+			// Standard Error: 9_449
+			.saturating_add(Weight::from_parts(144_837, 0).saturating_mul(m.into()))
+			// Standard Error: 2_528
+			.saturating_add(Weight::from_parts(390_566, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 1).saturating_mul(b.into()))
@@ -232,12 +232,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `61 + m * (64 ±0) + p * (54 ±0)`
 		//  Estimated: `4039 + m * (80 ±0) + p * (50 ±0)`
-		// Minimum execution time: 20_697_000 picoseconds.
-		Weight::from_parts(20_539_760, 4039)
-			// Standard Error: 5_939
-			.saturating_add(Weight::from_parts(164_806, 0).saturating_mul(m.into()))
-			// Standard Error: 1_584
-			.saturating_add(Weight::from_parts(231_075, 0).saturating_mul(p.into()))
+		// Minimum execution time: 20_249_000 picoseconds.
+		Weight::from_parts(20_211_402, 4039)
+			// Standard Error: 6_027
+			.saturating_add(Weight::from_parts(178_990, 0).saturating_mul(m.into()))
+			// Standard Error: 1_607
+			.saturating_add(Weight::from_parts(242_788, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 80).saturating_mul(m.into()))
@@ -263,14 +263,14 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `92 + b * (1 ±0) + m * (64 ±0) + p * (72 ±0)`
 		//  Estimated: `4045 + b * (1 ±0) + m * (85 ±0) + p * (65 ±0)`
-		// Minimum execution time: 29_950_000 picoseconds.
-		Weight::from_parts(30_954_629, 4045)
-			// Standard Error: 63
-			.saturating_add(Weight::from_parts(2_076, 0).saturating_mul(b.into()))
-			// Standard Error: 9_731
-			.saturating_add(Weight::from_parts(21_592, 0).saturating_mul(m.into()))
-			// Standard Error: 2_603
-			.saturating_add(Weight::from_parts(414_684, 0).saturating_mul(p.into()))
+		// Minimum execution time: 29_816_000 picoseconds.
+		Weight::from_parts(30_823_854, 4045)
+			// Standard Error: 71
+			.saturating_add(Weight::from_parts(2_028, 0).saturating_mul(b.into()))
+			// Standard Error: 11_088
+			.saturating_add(Weight::from_parts(33_208, 0).saturating_mul(m.into()))
+			// Standard Error: 2_966
+			.saturating_add(Weight::from_parts(430_100, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 1).saturating_mul(b.into()))
@@ -289,10 +289,10 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `226 + p * (32 ±0)`
 		//  Estimated: `2206 + p * (32 ±0)`
-		// Minimum execution time: 11_810_000 picoseconds.
-		Weight::from_parts(12_414_441, 2206)
-			// Standard Error: 1_001
-			.saturating_add(Weight::from_parts(201_207, 0).saturating_mul(p.into()))
+		// Minimum execution time: 11_321_000 picoseconds.
+		Weight::from_parts(11_931_897, 2206)
+			// Standard Error: 954
+			.saturating_add(Weight::from_parts(194_827, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(p.into()))
