@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_collective
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -50,13 +50,13 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 	fn set_members(m: u32, _n: u32, p: u32, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0 + m * (832 ±0) + p * (310 ±0)`
-		//  Estimated: `4856 + m * (489 ±5) + p * (2615 ±2)`
-		// Minimum execution time: 7_184_000 picoseconds.
-		Weight::from_parts(7_497_000, 4856)
-			// Standard Error: 62_768
-			.saturating_add(Weight::from_parts(3_005_906, 0).saturating_mul(m.into()))
-			// Standard Error: 25_506
-			.saturating_add(Weight::from_parts(3_381_038, 0).saturating_mul(p.into()))
+		//  Estimated: `4856 + m * (489 ±3) + p * (2615 ±1)`
+		// Minimum execution time: 7_177_000 picoseconds.
+		Weight::from_parts(7_449_000, 4856)
+			// Standard Error: 62_807
+			.saturating_add(Weight::from_parts(2_964_540, 0).saturating_mul(m.into()))
+			// Standard Error: 25_522
+			.saturating_add(Weight::from_parts(3_374_832, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(p.into())))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
@@ -74,12 +74,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `137 + m * (32 ±0)`
 		//  Estimated: `2116 + m * (32 ±0)`
-		// Minimum execution time: 11_059_000 picoseconds.
-		Weight::from_parts(11_161_392, 2116)
-			// Standard Error: 10
-			.saturating_add(Weight::from_parts(1_574, 0).saturating_mul(b.into()))
-			// Standard Error: 1_081
-			.saturating_add(Weight::from_parts(30_988, 0).saturating_mul(m.into()))
+		// Minimum execution time: 10_870_000 picoseconds.
+		Weight::from_parts(10_932_173, 2116)
+			// Standard Error: 9
+			.saturating_add(Weight::from_parts(1_494, 0).saturating_mul(b.into()))
+			// Standard Error: 980
+			.saturating_add(Weight::from_parts(26_280, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(m.into()))
 	}
@@ -95,12 +95,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `137 + m * (32 ±0)`
 		//  Estimated: `4096 + m * (32 ±0)`
-		// Minimum execution time: 13_506_000 picoseconds.
-		Weight::from_parts(13_572_454, 4096)
-			// Standard Error: 11
-			.saturating_add(Weight::from_parts(1_554, 0).saturating_mul(b.into()))
-			// Standard Error: 1_209
-			.saturating_add(Weight::from_parts(34_248, 0).saturating_mul(m.into()))
+		// Minimum execution time: 13_261_000 picoseconds.
+		Weight::from_parts(13_333_703, 4096)
+			// Standard Error: 10
+			.saturating_add(Weight::from_parts(1_559, 0).saturating_mul(b.into()))
+			// Standard Error: 1_167
+			.saturating_add(Weight::from_parts(32_432, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(m.into()))
 	}
@@ -124,14 +124,14 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `180 + m * (32 ±0) + p * (51 ±0)`
 		//  Estimated: `4092 + m * (42 ±0) + p * (49 ±0)`
-		// Minimum execution time: 18_261_000 picoseconds.
-		Weight::from_parts(17_032_141, 4092)
-			// Standard Error: 43
-			.saturating_add(Weight::from_parts(2_375, 0).saturating_mul(b.into()))
-			// Standard Error: 5_172
-			.saturating_add(Weight::from_parts(159_845, 0).saturating_mul(m.into()))
-			// Standard Error: 1_805
-			.saturating_add(Weight::from_parts(246_747, 0).saturating_mul(p.into()))
+		// Minimum execution time: 18_020_000 picoseconds.
+		Weight::from_parts(17_181_316, 4092)
+			// Standard Error: 46
+			.saturating_add(Weight::from_parts(2_065, 0).saturating_mul(b.into()))
+			// Standard Error: 5_532
+			.saturating_add(Weight::from_parts(149_144, 0).saturating_mul(m.into()))
+			// Standard Error: 1_930
+			.saturating_add(Weight::from_parts(257_427, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 			.saturating_add(Weight::from_parts(0, 42).saturating_mul(m.into()))
@@ -147,10 +147,10 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `638 + m * (64 ±0)`
 		//  Estimated: `4598 + m * (64 ±0)`
-		// Minimum execution time: 16_406_000 picoseconds.
-		Weight::from_parts(16_991_705, 4598)
-			// Standard Error: 4_046
-			.saturating_add(Weight::from_parts(75_770, 0).saturating_mul(m.into()))
+		// Minimum execution time: 15_927_000 picoseconds.
+		Weight::from_parts(16_561_031, 4598)
+			// Standard Error: 4_852
+			.saturating_add(Weight::from_parts(84_194, 0).saturating_mul(m.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 			.saturating_add(Weight::from_parts(0, 64).saturating_mul(m.into()))
@@ -171,12 +171,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `108 + m * (64 ±0) + p * (54 ±0)`
 		//  Estimated: `4086 + m * (80 ±0) + p * (50 ±0)`
-		// Minimum execution time: 18_935_000 picoseconds.
-		Weight::from_parts(18_953_627, 4086)
-			// Standard Error: 5_485
-			.saturating_add(Weight::from_parts(150_642, 0).saturating_mul(m.into()))
-			// Standard Error: 1_463
-			.saturating_add(Weight::from_parts(219_756, 0).saturating_mul(p.into()))
+		// Minimum execution time: 18_929_000 picoseconds.
+		Weight::from_parts(18_492_027, 4086)
+			// Standard Error: 4_715
+			.saturating_add(Weight::from_parts(169_942, 0).saturating_mul(m.into()))
+			// Standard Error: 1_257
+			.saturating_add(Weight::from_parts(244_497, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 80).saturating_mul(m.into()))
@@ -200,14 +200,14 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `139 + b * (1 ±0) + m * (64 ±0) + p * (72 ±0)`
 		//  Estimated: `4092 + b * (1 ±0) + m * (85 ±0) + p * (65 ±0)`
-		// Minimum execution time: 28_070_000 picoseconds.
-		Weight::from_parts(28_723_583, 4092)
-			// Standard Error: 62
-			.saturating_add(Weight::from_parts(1_838, 0).saturating_mul(b.into()))
-			// Standard Error: 9_608
-			.saturating_add(Weight::from_parts(109_978, 0).saturating_mul(m.into()))
-			// Standard Error: 2_570
-			.saturating_add(Weight::from_parts(369_670, 0).saturating_mul(p.into()))
+		// Minimum execution time: 28_111_000 picoseconds.
+		Weight::from_parts(28_041_589, 4092)
+			// Standard Error: 59
+			.saturating_add(Weight::from_parts(2_047, 0).saturating_mul(b.into()))
+			// Standard Error: 9_193
+			.saturating_add(Weight::from_parts(143_763, 0).saturating_mul(m.into()))
+			// Standard Error: 2_459
+			.saturating_add(Weight::from_parts(388_103, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 1).saturating_mul(b.into()))
@@ -232,12 +232,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `128 + m * (64 ±0) + p * (54 ±0)`
 		//  Estimated: `4106 + m * (80 ±0) + p * (50 ±0)`
-		// Minimum execution time: 20_725_000 picoseconds.
-		Weight::from_parts(20_846_243, 4106)
-			// Standard Error: 5_605
-			.saturating_add(Weight::from_parts(149_609, 0).saturating_mul(m.into()))
-			// Standard Error: 1_495
-			.saturating_add(Weight::from_parts(222_325, 0).saturating_mul(p.into()))
+		// Minimum execution time: 20_835_000 picoseconds.
+		Weight::from_parts(20_521_003, 4106)
+			// Standard Error: 5_358
+			.saturating_add(Weight::from_parts(163_779, 0).saturating_mul(m.into()))
+			// Standard Error: 1_429
+			.saturating_add(Weight::from_parts(243_154, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 80).saturating_mul(m.into()))
@@ -263,14 +263,12 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `159 + b * (1 ±0) + m * (64 ±0) + p * (72 ±0)`
 		//  Estimated: `4112 + b * (1 ±0) + m * (85 ±0) + p * (65 ±0)`
-		// Minimum execution time: 30_175_000 picoseconds.
-		Weight::from_parts(30_826_304, 4112)
-			// Standard Error: 60
-			.saturating_add(Weight::from_parts(2_007, 0).saturating_mul(b.into()))
-			// Standard Error: 9_350
-			.saturating_add(Weight::from_parts(55_681, 0).saturating_mul(m.into()))
-			// Standard Error: 2_501
-			.saturating_add(Weight::from_parts(401_849, 0).saturating_mul(p.into()))
+		// Minimum execution time: 30_178_000 picoseconds.
+		Weight::from_parts(31_500_786, 4112)
+			// Standard Error: 68
+			.saturating_add(Weight::from_parts(2_025, 0).saturating_mul(b.into()))
+			// Standard Error: 2_839
+			.saturating_add(Weight::from_parts(421_102, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 1).saturating_mul(b.into()))
@@ -289,10 +287,10 @@ impl<T: frame_system::Config> pallet_collective::WeightInfo for SubstrateWeight<
 		// Proof Size summary in bytes:
 		//  Measured:  `293 + p * (32 ±0)`
 		//  Estimated: `2273 + p * (32 ±0)`
-		// Minimum execution time: 11_845_000 picoseconds.
-		Weight::from_parts(12_552_807, 2273)
-			// Standard Error: 820
-			.saturating_add(Weight::from_parts(189_426, 0).saturating_mul(p.into()))
+		// Minimum execution time: 11_621_000 picoseconds.
+		Weight::from_parts(12_161_773, 2273)
+			// Standard Error: 860
+			.saturating_add(Weight::from_parts(190_367, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 			.saturating_add(Weight::from_parts(0, 32).saturating_mul(p.into()))
