@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_democracy
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -45,8 +45,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `4772`
 		//  Estimated: `18682`
-		// Minimum execution time: 39_317_000 picoseconds.
-		Weight::from_parts(40_991_000, 18682)
+		// Minimum execution time: 40_265_000 picoseconds.
+		Weight::from_parts(41_036_000, 18682)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -56,8 +56,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `3523`
 		//  Estimated: `7190`
-		// Minimum execution time: 35_103_000 picoseconds.
-		Weight::from_parts(36_586_000, 7190)
+		// Minimum execution time: 36_239_000 picoseconds.
+		Weight::from_parts(37_398_000, 7190)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -73,8 +73,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `3404`
 		//  Estimated: `7755`
-		// Minimum execution time: 53_753_000 picoseconds.
-		Weight::from_parts(55_224_000, 7755)
+		// Minimum execution time: 54_983_000 picoseconds.
+		Weight::from_parts(56_207_000, 7755)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -90,8 +90,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `3426`
 		//  Estimated: `7755`
-		// Minimum execution time: 57_356_000 picoseconds.
-		Weight::from_parts(58_364_000, 7755)
+		// Minimum execution time: 57_588_000 picoseconds.
+		Weight::from_parts(59_018_000, 7755)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -105,8 +105,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `299`
 		//  Estimated: `4161`
-		// Minimum execution time: 21_235_000 picoseconds.
-		Weight::from_parts(21_570_000, 4161)
+		// Minimum execution time: 21_115_000 picoseconds.
+		Weight::from_parts(21_499_000, 4161)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -128,8 +128,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `5942`
 		//  Estimated: `18682`
-		// Minimum execution time: 84_108_000 picoseconds.
-		Weight::from_parts(86_640_000, 18682)
+		// Minimum execution time: 85_918_000 picoseconds.
+		Weight::from_parts(88_147_000, 18682)
 			.saturating_add(T::DbWeight::get().reads(8_u64))
 			.saturating_add(T::DbWeight::get().writes(7_u64))
 	}
@@ -141,8 +141,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `3349`
 		//  Estimated: `7198`
-		// Minimum execution time: 10_968_000 picoseconds.
-		Weight::from_parts(11_435_000, 7198)
+		// Minimum execution time: 10_962_000 picoseconds.
+		Weight::from_parts(11_592_000, 7198)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -152,8 +152,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 2_240_000 picoseconds.
-		Weight::from_parts(2_430_000, 0)
+		// Minimum execution time: 2_133_000 picoseconds.
+		Weight::from_parts(2_305_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	/// Storage: `Democracy::NextExternal` (r:0 w:1)
@@ -162,8 +162,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 2_264_000 picoseconds.
-		Weight::from_parts(2_505_000, 0)
+		// Minimum execution time: 2_175_000 picoseconds.
+		Weight::from_parts(2_308_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	/// Storage: `Democracy::NextExternal` (r:1 w:1)
@@ -178,8 +178,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `219`
 		//  Estimated: `4013`
-		// Minimum execution time: 20_592_000 picoseconds.
-		Weight::from_parts(21_137_000, 4013)
+		// Minimum execution time: 19_893_000 picoseconds.
+		Weight::from_parts(20_613_000, 4013)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(5_u64))
 	}
@@ -193,8 +193,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `3452`
 		//  Estimated: `7198`
-		// Minimum execution time: 23_603_000 picoseconds.
-		Weight::from_parts(24_386_000, 7198)
+		// Minimum execution time: 23_467_000 picoseconds.
+		Weight::from_parts(24_029_000, 7198)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -210,8 +210,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `5853`
 		//  Estimated: `18682`
-		// Minimum execution time: 68_197_000 picoseconds.
-		Weight::from_parts(69_538_000, 18682)
+		// Minimum execution time: 70_224_000 picoseconds.
+		Weight::from_parts(71_096_000, 18682)
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
@@ -223,8 +223,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `204`
 		//  Estimated: `4013`
-		// Minimum execution time: 14_672_000 picoseconds.
-		Weight::from_parts(14_941_000, 4013)
+		// Minimum execution time: 14_291_000 picoseconds.
+		Weight::from_parts(14_687_000, 4013)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -239,10 +239,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `177 + r * (86 ±0)`
 		//  Estimated: `1984 + r * (2676 ±0)`
-		// Minimum execution time: 5_112_000 picoseconds.
-		Weight::from_parts(5_878_095, 1984)
-			// Standard Error: 7_412
-			.saturating_add(Weight::from_parts(3_124_027, 0).saturating_mul(r.into()))
+		// Minimum execution time: 5_080_000 picoseconds.
+		Weight::from_parts(6_537_378, 1984)
+			// Standard Error: 6_384
+			.saturating_add(Weight::from_parts(3_106_919, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(r.into())))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
@@ -265,10 +265,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `177 + r * (86 ±0)`
 		//  Estimated: `18682 + r * (2676 ±0)`
-		// Minimum execution time: 7_147_000 picoseconds.
-		Weight::from_parts(8_738_111, 18682)
-			// Standard Error: 6_333
-			.saturating_add(Weight::from_parts(3_124_021, 0).saturating_mul(r.into()))
+		// Minimum execution time: 7_318_000 picoseconds.
+		Weight::from_parts(7_735_718, 18682)
+			// Standard Error: 6_602
+			.saturating_add(Weight::from_parts(3_171_160, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(r.into())))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
@@ -287,10 +287,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `768 + r * (108 ±0)`
 		//  Estimated: `20295 + r * (2676 ±0)`
-		// Minimum execution time: 38_544_000 picoseconds.
-		Weight::from_parts(41_814_208, 20295)
-			// Standard Error: 7_825
-			.saturating_add(Weight::from_parts(4_029_241, 0).saturating_mul(r.into()))
+		// Minimum execution time: 39_300_000 picoseconds.
+		Weight::from_parts(43_023_454, 20295)
+			// Standard Error: 7_400
+			.saturating_add(Weight::from_parts(3_980_244, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(5_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(r.into())))
 			.saturating_add(T::DbWeight::get().writes(4_u64))
@@ -306,10 +306,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `426 + r * (108 ±0)`
 		//  Estimated: `14025 + r * (2676 ±0)`
-		// Minimum execution time: 17_243_000 picoseconds.
-		Weight::from_parts(17_694_932, 14025)
-			// Standard Error: 8_180
-			.saturating_add(Weight::from_parts(3_988_312, 0).saturating_mul(r.into()))
+		// Minimum execution time: 16_707_000 picoseconds.
+		Weight::from_parts(18_559_244, 14025)
+			// Standard Error: 5_909
+			.saturating_add(Weight::from_parts(3_907_228, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().reads((1_u64).saturating_mul(r.into())))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
@@ -322,8 +322,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 2_356_000 picoseconds.
-		Weight::from_parts(2_498_000, 0)
+		// Minimum execution time: 2_430_000 picoseconds.
+		Weight::from_parts(2_552_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	/// Storage: `Democracy::VotingOf` (r:1 w:1)
@@ -339,10 +339,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `496`
 		//  Estimated: `7755`
-		// Minimum execution time: 21_832_000 picoseconds.
-		Weight::from_parts(33_058_579, 7755)
-			// Standard Error: 2_994
-			.saturating_add(Weight::from_parts(91_740, 0).saturating_mul(r.into()))
+		// Minimum execution time: 21_693_000 picoseconds.
+		Weight::from_parts(32_923_984, 7755)
+			// Standard Error: 2_834
+			.saturating_add(Weight::from_parts(90_150, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -359,10 +359,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `497 + r * (22 ±0)`
 		//  Estimated: `7755`
-		// Minimum execution time: 29_512_000 picoseconds.
-		Weight::from_parts(32_520_107, 7755)
-			// Standard Error: 1_827
-			.saturating_add(Weight::from_parts(128_464, 0).saturating_mul(r.into()))
+		// Minimum execution time: 30_784_000 picoseconds.
+		Weight::from_parts(33_892_263, 7755)
+			// Standard Error: 1_682
+			.saturating_add(Weight::from_parts(114_424, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(4_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -375,10 +375,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `661 + r * (26 ±0)`
 		//  Estimated: `7755`
-		// Minimum execution time: 15_022_000 picoseconds.
-		Weight::from_parts(17_683_943, 7755)
-			// Standard Error: 1_671
-			.saturating_add(Weight::from_parts(119_222, 0).saturating_mul(r.into()))
+		// Minimum execution time: 14_607_000 picoseconds.
+		Weight::from_parts(17_663_577, 7755)
+			// Standard Error: 1_751
+			.saturating_add(Weight::from_parts(110_997, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -391,10 +391,10 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `661 + r * (26 ±0)`
 		//  Estimated: `7755`
-		// Minimum execution time: 14_724_000 picoseconds.
-		Weight::from_parts(17_618_848, 7755)
-			// Standard Error: 1_733
-			.saturating_add(Weight::from_parts(125_636, 0).saturating_mul(r.into()))
+		// Minimum execution time: 14_607_000 picoseconds.
+		Weight::from_parts(17_872_390, 7755)
+			// Standard Error: 1_776
+			.saturating_add(Weight::from_parts(111_908, 0).saturating_mul(r.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -410,8 +410,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `289`
 		//  Estimated: `4051`
-		// Minimum execution time: 16_030_000 picoseconds.
-		Weight::from_parts(16_502_000, 4051)
+		// Minimum execution time: 15_620_000 picoseconds.
+		Weight::from_parts(16_201_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -423,8 +423,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `219`
 		//  Estimated: `4013`
-		// Minimum execution time: 13_044_000 picoseconds.
-		Weight::from_parts(13_383_000, 4013)
+		// Minimum execution time: 12_960_000 picoseconds.
+		Weight::from_parts(13_141_000, 4013)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -440,8 +440,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `4821`
 		//  Estimated: `18682`
-		// Minimum execution time: 40_286_000 picoseconds.
-		Weight::from_parts(41_294_000, 18682)
+		// Minimum execution time: 39_318_000 picoseconds.
+		Weight::from_parts(39_971_000, 18682)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -453,8 +453,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `4755`
 		//  Estimated: `18682`
-		// Minimum execution time: 34_782_000 picoseconds.
-		Weight::from_parts(35_546_000, 18682)
+		// Minimum execution time: 34_884_000 picoseconds.
+		Weight::from_parts(35_874_000, 18682)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -468,8 +468,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `144`
 		//  Estimated: `4051`
-		// Minimum execution time: 12_911_000 picoseconds.
-		Weight::from_parts(13_257_000, 4051)
+		// Minimum execution time: 12_843_000 picoseconds.
+		Weight::from_parts(13_024_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -481,8 +481,8 @@ impl<T: frame_system::Config> pallet_democracy::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `235`
 		//  Estimated: `4161`
-		// Minimum execution time: 15_643_000 picoseconds.
-		Weight::from_parts(16_137_000, 4161)
+		// Minimum execution time: 15_168_000 picoseconds.
+		Weight::from_parts(15_474_000, 4161)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
