@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_multisig
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -38,10 +38,10 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 12_626_000 picoseconds.
-		Weight::from_parts(13_369_367, 0)
-			// Standard Error: 3
-			.saturating_add(Weight::from_parts(544, 0).saturating_mul(z.into()))
+		// Minimum execution time: 12_451_000 picoseconds.
+		Weight::from_parts(13_046_686, 0)
+			// Standard Error: 2
+			.saturating_add(Weight::from_parts(553, 0).saturating_mul(z.into()))
 	}
 	/// Storage: `Multisig::Multisigs` (r:1 w:1)
 	/// Proof: `Multisig::Multisigs` (`max_values`: None, `max_size`: Some(3346), added: 5821, mode: `MaxEncodedLen`)
@@ -51,12 +51,12 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `192 + s * (2 ±0)`
 		//  Estimated: `7306`
-		// Minimum execution time: 37_677_000 picoseconds.
-		Weight::from_parts(27_324_554, 7306)
-			// Standard Error: 528
-			.saturating_add(Weight::from_parts(114_718, 0).saturating_mul(s.into()))
-			// Standard Error: 5
-			.saturating_add(Weight::from_parts(1_468, 0).saturating_mul(z.into()))
+		// Minimum execution time: 38_932_000 picoseconds.
+		Weight::from_parts(29_320_977, 7306)
+			// Standard Error: 509
+			.saturating_add(Weight::from_parts(108_143, 0).saturating_mul(s.into()))
+			// Standard Error: 4
+			.saturating_add(Weight::from_parts(1_459, 0).saturating_mul(z.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -68,12 +68,12 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `211`
 		//  Estimated: `7306`
-		// Minimum execution time: 24_801_000 picoseconds.
-		Weight::from_parts(15_332_807, 7306)
-			// Standard Error: 321
-			.saturating_add(Weight::from_parts(105_667, 0).saturating_mul(s.into()))
+		// Minimum execution time: 24_378_000 picoseconds.
+		Weight::from_parts(15_575_660, 7306)
+			// Standard Error: 362
+			.saturating_add(Weight::from_parts(97_278, 0).saturating_mul(s.into()))
 			// Standard Error: 3
-			.saturating_add(Weight::from_parts(1_444, 0).saturating_mul(z.into()))
+			.saturating_add(Weight::from_parts(1_476, 0).saturating_mul(z.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -87,12 +87,12 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `314 + s * (33 ±0)`
 		//  Estimated: `7306`
-		// Minimum execution time: 43_929_000 picoseconds.
-		Weight::from_parts(29_894_823, 7306)
-			// Standard Error: 714
-			.saturating_add(Weight::from_parts(151_210, 0).saturating_mul(s.into()))
-			// Standard Error: 7
-			.saturating_add(Weight::from_parts(1_569, 0).saturating_mul(z.into()))
+		// Minimum execution time: 43_558_000 picoseconds.
+		Weight::from_parts(31_499_024, 7306)
+			// Standard Error: 556
+			.saturating_add(Weight::from_parts(138_872, 0).saturating_mul(s.into()))
+			// Standard Error: 5
+			.saturating_add(Weight::from_parts(1_505, 0).saturating_mul(z.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -103,10 +103,10 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `193 + s * (2 ±0)`
 		//  Estimated: `7306`
-		// Minimum execution time: 25_026_000 picoseconds.
-		Weight::from_parts(26_284_298, 7306)
-			// Standard Error: 610
-			.saturating_add(Weight::from_parts(115_884, 0).saturating_mul(s.into()))
+		// Minimum execution time: 26_602_000 picoseconds.
+		Weight::from_parts(27_813_289, 7306)
+			// Standard Error: 604
+			.saturating_add(Weight::from_parts(112_864, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -117,10 +117,10 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `211`
 		//  Estimated: `7306`
-		// Minimum execution time: 13_760_000 picoseconds.
-		Weight::from_parts(14_220_855, 7306)
-			// Standard Error: 417
-			.saturating_add(Weight::from_parts(102_769, 0).saturating_mul(s.into()))
+		// Minimum execution time: 13_876_000 picoseconds.
+		Weight::from_parts(14_331_399, 7306)
+			// Standard Error: 443
+			.saturating_add(Weight::from_parts(102_534, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -131,10 +131,10 @@ impl<T: frame_system::Config> pallet_multisig::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `383 + s * (1 ±0)`
 		//  Estimated: `7306`
-		// Minimum execution time: 26_341_000 picoseconds.
-		Weight::from_parts(27_351_588, 7306)
-			// Standard Error: 546
-			.saturating_add(Weight::from_parts(113_945, 0).saturating_mul(s.into()))
+		// Minimum execution time: 26_279_000 picoseconds.
+		Weight::from_parts(27_481_125, 7306)
+			// Standard Error: 592
+			.saturating_add(Weight::from_parts(113_118, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
