@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_preimage
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -46,10 +46,10 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `42`
 		//  Estimated: `4051`
-		// Minimum execution time: 43_232_000 picoseconds.
-		Weight::from_parts(43_545_000, 4051)
-			// Standard Error: 11
-			.saturating_add(Weight::from_parts(2_596, 0).saturating_mul(s.into()))
+		// Minimum execution time: 43_465_000 picoseconds.
+		Weight::from_parts(43_893_000, 4051)
+			// Standard Error: 10
+			.saturating_add(Weight::from_parts(2_454, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -64,10 +64,10 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `106`
 		//  Estimated: `4051`
-		// Minimum execution time: 14_098_000 picoseconds.
-		Weight::from_parts(14_237_000, 4051)
-			// Standard Error: 10
-			.saturating_add(Weight::from_parts(2_591, 0).saturating_mul(s.into()))
+		// Minimum execution time: 13_781_000 picoseconds.
+		Weight::from_parts(14_099_000, 4051)
+			// Standard Error: 8
+			.saturating_add(Weight::from_parts(2_461, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -82,10 +82,10 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `106`
 		//  Estimated: `4051`
-		// Minimum execution time: 13_230_000 picoseconds.
-		Weight::from_parts(13_440_000, 4051)
-			// Standard Error: 10
-			.saturating_add(Weight::from_parts(2_607, 0).saturating_mul(s.into()))
+		// Minimum execution time: 13_053_000 picoseconds.
+		Weight::from_parts(13_327_000, 4051)
+			// Standard Error: 8
+			.saturating_add(Weight::from_parts(2_442, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -101,8 +101,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `243`
 		//  Estimated: `4051`
-		// Minimum execution time: 69_658_000 picoseconds.
-		Weight::from_parts(76_033_000, 4051)
+		// Minimum execution time: 46_459_000 picoseconds.
+		Weight::from_parts(47_760_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -116,8 +116,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `144`
 		//  Estimated: `4051`
-		// Minimum execution time: 29_715_000 picoseconds.
-		Weight::from_parts(35_827_000, 4051)
+		// Minimum execution time: 21_187_000 picoseconds.
+		Weight::from_parts(21_587_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -129,8 +129,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `188`
 		//  Estimated: `4051`
-		// Minimum execution time: 28_760_000 picoseconds.
-		Weight::from_parts(32_352_000, 4051)
+		// Minimum execution time: 17_881_000 picoseconds.
+		Weight::from_parts(18_237_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -142,8 +142,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `144`
 		//  Estimated: `4051`
-		// Minimum execution time: 23_630_000 picoseconds.
-		Weight::from_parts(25_878_000, 4051)
+		// Minimum execution time: 13_047_000 picoseconds.
+		Weight::from_parts(13_478_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -155,8 +155,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `42`
 		//  Estimated: `4051`
-		// Minimum execution time: 19_375_000 picoseconds.
-		Weight::from_parts(25_208_000, 4051)
+		// Minimum execution time: 13_968_000 picoseconds.
+		Weight::from_parts(14_664_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -168,8 +168,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `106`
 		//  Estimated: `4051`
-		// Minimum execution time: 12_140_000 picoseconds.
-		Weight::from_parts(12_723_000, 4051)
+		// Minimum execution time: 9_503_000 picoseconds.
+		Weight::from_parts(9_692_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -183,8 +183,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `144`
 		//  Estimated: `4051`
-		// Minimum execution time: 29_975_000 picoseconds.
-		Weight::from_parts(33_764_000, 4051)
+		// Minimum execution time: 18_862_000 picoseconds.
+		Weight::from_parts(19_185_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -196,8 +196,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `106`
 		//  Estimated: `4051`
-		// Minimum execution time: 11_927_000 picoseconds.
-		Weight::from_parts(12_778_000, 4051)
+		// Minimum execution time: 9_613_000 picoseconds.
+		Weight::from_parts(9_856_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -209,8 +209,8 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `106`
 		//  Estimated: `4051`
-		// Minimum execution time: 11_102_000 picoseconds.
-		Weight::from_parts(12_362_000, 4051)
+		// Minimum execution time: 9_105_000 picoseconds.
+		Weight::from_parts(9_481_000, 4051)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -227,10 +227,10 @@ impl<T: frame_system::Config> pallet_preimage::WeightInfo for SubstrateWeight<T>
 		// Proof Size summary in bytes:
 		//  Measured:  `233 + n * (227 ±0)`
 		//  Estimated: `1485 + n * (2603 ±0)`
-		// Minimum execution time: 48_939_000 picoseconds.
-		Weight::from_parts(49_467_000, 1485)
-			// Standard Error: 60_059
-			.saturating_add(Weight::from_parts(50_936_619, 0).saturating_mul(n.into()))
+		// Minimum execution time: 48_802_000 picoseconds.
+		Weight::from_parts(49_311_000, 1485)
+			// Standard Error: 30_390
+			.saturating_add(Weight::from_parts(49_227_432, 0).saturating_mul(n.into()))
 			.saturating_add(T::DbWeight::get().reads((3_u64).saturating_mul(n.into())))
 			.saturating_add(T::DbWeight::get().writes((4_u64).saturating_mul(n.into())))
 			.saturating_add(Weight::from_parts(0, 2603).saturating_mul(n.into()))
