@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_proxy
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -40,10 +40,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161 + p * (37 ±0)`
 		//  Estimated: `5201`
-		// Minimum execution time: 12_200_000 picoseconds.
-		Weight::from_parts(12_799_250, 5201)
-			// Standard Error: 821
-			.saturating_add(Weight::from_parts(45_911, 0).saturating_mul(p.into()))
+		// Minimum execution time: 12_059_000 picoseconds.
+		Weight::from_parts(12_597_553, 5201)
+			// Standard Error: 919
+			.saturating_add(Weight::from_parts(42_369, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 	}
 	/// Storage: `Proxy::Proxies` (r:1 w:0)
@@ -58,12 +58,12 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `524 + a * (68 ±0) + p * (37 ±0)`
 		//  Estimated: `6193`
-		// Minimum execution time: 34_650_000 picoseconds.
-		Weight::from_parts(33_134_945, 6193)
-			// Standard Error: 4_701
-			.saturating_add(Weight::from_parts(123_021, 0).saturating_mul(a.into()))
-			// Standard Error: 4_857
-			.saturating_add(Weight::from_parts(112_627, 0).saturating_mul(p.into()))
+		// Minimum execution time: 35_173_000 picoseconds.
+		Weight::from_parts(33_695_300, 6193)
+			// Standard Error: 4_373
+			.saturating_add(Weight::from_parts(121_953, 0).saturating_mul(a.into()))
+			// Standard Error: 4_519
+			.saturating_add(Weight::from_parts(109_879, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -77,12 +77,12 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `441 + a * (68 ±0)`
 		//  Estimated: `6193`
-		// Minimum execution time: 21_867_000 picoseconds.
-		Weight::from_parts(21_796_650, 6193)
-			// Standard Error: 1_507
-			.saturating_add(Weight::from_parts(134_168, 0).saturating_mul(a.into()))
-			// Standard Error: 1_557
-			.saturating_add(Weight::from_parts(34_824, 0).saturating_mul(p.into()))
+		// Minimum execution time: 22_572_000 picoseconds.
+		Weight::from_parts(22_318_291, 6193)
+			// Standard Error: 1_513
+			.saturating_add(Weight::from_parts(128_593, 0).saturating_mul(a.into()))
+			// Standard Error: 1_563
+			.saturating_add(Weight::from_parts(39_305, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -96,12 +96,12 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `441 + a * (68 ±0)`
 		//  Estimated: `6193`
-		// Minimum execution time: 21_901_000 picoseconds.
-		Weight::from_parts(21_740_015, 6193)
-			// Standard Error: 1_501
-			.saturating_add(Weight::from_parts(133_319, 0).saturating_mul(a.into()))
-			// Standard Error: 1_551
-			.saturating_add(Weight::from_parts(35_852, 0).saturating_mul(p.into()))
+		// Minimum execution time: 22_310_000 picoseconds.
+		Weight::from_parts(22_211_872, 6193)
+			// Standard Error: 1_437
+			.saturating_add(Weight::from_parts(129_022, 0).saturating_mul(a.into()))
+			// Standard Error: 1_484
+			.saturating_add(Weight::from_parts(39_826, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -117,12 +117,12 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `456 + a * (68 ±0) + p * (37 ±0)`
 		//  Estimated: `6193`
-		// Minimum execution time: 28_907_000 picoseconds.
-		Weight::from_parts(28_918_368, 6193)
-			// Standard Error: 4_598
-			.saturating_add(Weight::from_parts(132_396, 0).saturating_mul(a.into()))
-			// Standard Error: 4_750
-			.saturating_add(Weight::from_parts(114_773, 0).saturating_mul(p.into()))
+		// Minimum execution time: 29_906_000 picoseconds.
+		Weight::from_parts(29_692_239, 6193)
+			// Standard Error: 4_156
+			.saturating_add(Weight::from_parts(126_544, 0).saturating_mul(a.into()))
+			// Standard Error: 4_294
+			.saturating_add(Weight::from_parts(110_567, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -133,10 +133,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161 + p * (37 ±0)`
 		//  Estimated: `5201`
-		// Minimum execution time: 19_614_000 picoseconds.
-		Weight::from_parts(20_338_700, 5201)
-			// Standard Error: 922
-			.saturating_add(Weight::from_parts(39_235, 0).saturating_mul(p.into()))
+		// Minimum execution time: 20_652_000 picoseconds.
+		Weight::from_parts(21_399_005, 5201)
+			// Standard Error: 1_031
+			.saturating_add(Weight::from_parts(36_414, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -147,10 +147,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161 + p * (37 ±0)`
 		//  Estimated: `5201`
-		// Minimum execution time: 19_437_000 picoseconds.
-		Weight::from_parts(20_627_246, 5201)
-			// Standard Error: 1_627
-			.saturating_add(Weight::from_parts(37_025, 0).saturating_mul(p.into()))
+		// Minimum execution time: 20_632_000 picoseconds.
+		Weight::from_parts(21_378_187, 5201)
+			// Standard Error: 1_063
+			.saturating_add(Weight::from_parts(52_076, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -161,10 +161,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `161 + p * (37 ±0)`
 		//  Estimated: `5201`
-		// Minimum execution time: 18_851_000 picoseconds.
-		Weight::from_parts(19_805_046, 5201)
-			// Standard Error: 1_156
-			.saturating_add(Weight::from_parts(31_325, 0).saturating_mul(p.into()))
+		// Minimum execution time: 18_915_000 picoseconds.
+		Weight::from_parts(19_652_423, 5201)
+			// Standard Error: 1_001
+			.saturating_add(Weight::from_parts(31_438, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -175,10 +175,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `173`
 		//  Estimated: `5201`
-		// Minimum execution time: 20_905_000 picoseconds.
-		Weight::from_parts(21_713_385, 5201)
-			// Standard Error: 1_016
-			.saturating_add(Weight::from_parts(10_006, 0).saturating_mul(p.into()))
+		// Minimum execution time: 22_133_000 picoseconds.
+		Weight::from_parts(22_888_148, 5201)
+			// Standard Error: 1_189
+			.saturating_add(Weight::from_parts(10_279, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -189,10 +189,10 @@ impl<T: frame_system::Config> pallet_proxy::WeightInfo for SubstrateWeight<T> {
 		// Proof Size summary in bytes:
 		//  Measured:  `198 + p * (37 ±0)`
 		//  Estimated: `5201`
-		// Minimum execution time: 19_767_000 picoseconds.
-		Weight::from_parts(20_759_938, 5201)
-			// Standard Error: 1_084
-			.saturating_add(Weight::from_parts(36_033, 0).saturating_mul(p.into()))
+		// Minimum execution time: 19_789_000 picoseconds.
+		Weight::from_parts(20_729_634, 5201)
+			// Standard Error: 995
+			.saturating_add(Weight::from_parts(34_157, 0).saturating_mul(p.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
