@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_scheduler
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -39,8 +39,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `31`
 		//  Estimated: `1984`
-		// Minimum execution time: 2_842_000 picoseconds.
-		Weight::from_parts(2_974_000, 1984)
+		// Minimum execution time: 2_661_000 picoseconds.
+		Weight::from_parts(2_834_000, 1984)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -51,10 +51,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `78 + s * (177 ±0)`
 		//  Estimated: `14423`
-		// Minimum execution time: 3_520_000 picoseconds.
-		Weight::from_parts(6_092_038, 14423)
-			// Standard Error: 1_341
-			.saturating_add(Weight::from_parts(404_544, 0).saturating_mul(s.into()))
+		// Minimum execution time: 3_283_000 picoseconds.
+		Weight::from_parts(5_744_775, 14423)
+			// Standard Error: 1_255
+			.saturating_add(Weight::from_parts(406_806, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -62,8 +62,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 2_540_000 picoseconds.
-		Weight::from_parts(2_688_000, 0)
+		// Minimum execution time: 2_375_000 picoseconds.
+		Weight::from_parts(2_531_000, 0)
 	}
 	/// Storage: `Preimage::PreimageFor` (r:1 w:1)
 	/// Proof: `Preimage::PreimageFor` (`max_values`: None, `max_size`: Some(4194344), added: 4196819, mode: `Measured`)
@@ -76,10 +76,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `179 + s * (1 ±0)`
 		//  Estimated: `4139 + s * (1 ±0)`
-		// Minimum execution time: 15_918_000 picoseconds.
-		Weight::from_parts(16_103_000, 4139)
+		// Minimum execution time: 15_654_000 picoseconds.
+		Weight::from_parts(15_839_000, 4139)
 			// Standard Error: 10
-			.saturating_add(Weight::from_parts(1_279, 0).saturating_mul(s.into()))
+			.saturating_add(Weight::from_parts(1_337, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(3_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 			.saturating_add(Weight::from_parts(0, 1).saturating_mul(s.into()))
@@ -90,30 +90,30 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_820_000 picoseconds.
-		Weight::from_parts(4_058_000, 0)
+		// Minimum execution time: 3_687_000 picoseconds.
+		Weight::from_parts(3_875_000, 0)
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	fn service_task_periodic() -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 2_589_000 picoseconds.
-		Weight::from_parts(2_727_000, 0)
+		// Minimum execution time: 2_362_000 picoseconds.
+		Weight::from_parts(2_508_000, 0)
 	}
 	fn execute_dispatch_signed() -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 1_842_000 picoseconds.
-		Weight::from_parts(2_002_000, 0)
+		// Minimum execution time: 1_863_000 picoseconds.
+		Weight::from_parts(1_968_000, 0)
 	}
 	fn execute_dispatch_unsigned() -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 1_784_000 picoseconds.
-		Weight::from_parts(1_930_000, 0)
+		// Minimum execution time: 1_861_000 picoseconds.
+		Weight::from_parts(1_942_000, 0)
 	}
 	/// Storage: `Scheduler::Agenda` (r:1 w:1)
 	/// Proof: `Scheduler::Agenda` (`max_values`: None, `max_size`: Some(10463), added: 12938, mode: `MaxEncodedLen`)
@@ -122,10 +122,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `78 + s * (177 ±0)`
 		//  Estimated: `14423`
-		// Minimum execution time: 8_939_000 picoseconds.
-		Weight::from_parts(11_956_545, 14423)
-			// Standard Error: 1_493
-			.saturating_add(Weight::from_parts(428_356, 0).saturating_mul(s.into()))
+		// Minimum execution time: 8_803_000 picoseconds.
+		Weight::from_parts(11_666_980, 14423)
+			// Standard Error: 1_485
+			.saturating_add(Weight::from_parts(424_352, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -140,10 +140,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `78 + s * (177 ±0)`
 		//  Estimated: `14423`
-		// Minimum execution time: 14_929_000 picoseconds.
-		Weight::from_parts(14_446_284, 14423)
-			// Standard Error: 958
-			.saturating_add(Weight::from_parts(655_563, 0).saturating_mul(s.into()))
+		// Minimum execution time: 14_409_000 picoseconds.
+		Weight::from_parts(13_978_199, 14423)
+			// Standard Error: 999
+			.saturating_add(Weight::from_parts(657_242, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -156,10 +156,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `255 + s * (185 ±0)`
 		//  Estimated: `14423`
-		// Minimum execution time: 12_334_000 picoseconds.
-		Weight::from_parts(15_920_861, 14423)
-			// Standard Error: 2_179
-			.saturating_add(Weight::from_parts(460_252, 0).saturating_mul(s.into()))
+		// Minimum execution time: 12_043_000 picoseconds.
+		Weight::from_parts(15_558_395, 14423)
+			// Standard Error: 2_271
+			.saturating_add(Weight::from_parts(453_181, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -174,10 +174,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `281 + s * (185 ±0)`
 		//  Estimated: `14423`
-		// Minimum execution time: 16_681_000 picoseconds.
-		Weight::from_parts(16_956_414, 14423)
-			// Standard Error: 1_225
-			.saturating_add(Weight::from_parts(690_218, 0).saturating_mul(s.into()))
+		// Minimum execution time: 16_502_000 picoseconds.
+		Weight::from_parts(16_562_361, 14423)
+			// Standard Error: 1_236
+			.saturating_add(Weight::from_parts(678_973, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
@@ -190,10 +190,10 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `118`
 		//  Estimated: `14423`
-		// Minimum execution time: 8_438_000 picoseconds.
-		Weight::from_parts(8_870_412, 14423)
-			// Standard Error: 397
-			.saturating_add(Weight::from_parts(23_161, 0).saturating_mul(s.into()))
+		// Minimum execution time: 8_292_000 picoseconds.
+		Weight::from_parts(8_637_034, 14423)
+			// Standard Error: 455
+			.saturating_add(Weight::from_parts(29_450, 0).saturating_mul(s.into()))
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
@@ -205,8 +205,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `8928`
 		//  Estimated: `14423`
-		// Minimum execution time: 24_739_000 picoseconds.
-		Weight::from_parts(25_815_000, 14423)
+		// Minimum execution time: 24_785_000 picoseconds.
+		Weight::from_parts(25_779_000, 14423)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -220,8 +220,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `9606`
 		//  Estimated: `14423`
-		// Minimum execution time: 30_409_000 picoseconds.
-		Weight::from_parts(31_267_000, 14423)
+		// Minimum execution time: 29_779_000 picoseconds.
+		Weight::from_parts(30_727_000, 14423)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -233,8 +233,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `8940`
 		//  Estimated: `14423`
-		// Minimum execution time: 24_147_000 picoseconds.
-		Weight::from_parts(24_660_000, 14423)
+		// Minimum execution time: 23_708_000 picoseconds.
+		Weight::from_parts(24_677_000, 14423)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -248,8 +248,8 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `9618`
 		//  Estimated: `14423`
-		// Minimum execution time: 29_347_000 picoseconds.
-		Weight::from_parts(30_286_000, 14423)
+		// Minimum execution time: 28_807_000 picoseconds.
+		Weight::from_parts(29_501_000, 14423)
 			.saturating_add(T::DbWeight::get().reads(2_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
