@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_timestamp
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -39,8 +39,8 @@ impl<T: frame_system::Config> pallet_timestamp::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `42`
 		//  Estimated: `1988`
-		// Minimum execution time: 4_510_000 picoseconds.
-		Weight::from_parts(4_827_000, 1988)
+		// Minimum execution time: 4_536_000 picoseconds.
+		Weight::from_parts(4_729_000, 1988)
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
@@ -48,8 +48,8 @@ impl<T: frame_system::Config> pallet_timestamp::WeightInfo for SubstrateWeight<T
 		// Proof Size summary in bytes:
 		//  Measured:  `94`
 		//  Estimated: `0`
-		// Minimum execution time: 3_414_000 picoseconds.
-		Weight::from_parts(3_603_000, 0)
+		// Minimum execution time: 3_394_000 picoseconds.
+		Weight::from_parts(3_517_000, 0)
 	}
 }
 
