@@ -1,9 +1,9 @@
 //! Autogenerated weights for pallet_utility
 //!
 //! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 32.0.0
-//! DATE: 2024-08-15, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
+//! DATE: 2024-09-04, STEPS: `50`, REPEAT: `20`, LOW RANGE: `[]`, HIGH RANGE: `[]`
 //! WORST CASE MAP SIZE: `1000000`
-//! HOSTNAME: `ip-10-173-11-47`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
+//! HOSTNAME: `ip-10-173-10-33`, CPU: `Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz`
 //! EXECUTION: , WASM-EXECUTION: Compiled, CHAIN: Some("frequency-bench"), DB CACHE: 1024
 
 // Executed Command:
@@ -38,44 +38,44 @@ impl<T: frame_system::Config> pallet_utility::WeightInfo for SubstrateWeight<T> 
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_871_000 picoseconds.
-		Weight::from_parts(7_515_194, 0)
-			// Standard Error: 1_080
-			.saturating_add(Weight::from_parts(2_715_492, 0).saturating_mul(c.into()))
+		// Minimum execution time: 3_694_000 picoseconds.
+		Weight::from_parts(5_422_646, 0)
+			// Standard Error: 863
+			.saturating_add(Weight::from_parts(2_812_016, 0).saturating_mul(c.into()))
 	}
 	fn as_derivative() -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_797_000 picoseconds.
-		Weight::from_parts(4_020_000, 0)
+		// Minimum execution time: 3_835_000 picoseconds.
+		Weight::from_parts(3_949_000, 0)
 	}
 	/// The range of component `c` is `[0, 1000]`.
 	fn batch_all(c: u32, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_830_000 picoseconds.
-		Weight::from_parts(3_942_838, 0)
-			// Standard Error: 1_296
-			.saturating_add(Weight::from_parts(2_966_382, 0).saturating_mul(c.into()))
+		// Minimum execution time: 3_698_000 picoseconds.
+		Weight::from_parts(2_862_085, 0)
+			// Standard Error: 887
+			.saturating_add(Weight::from_parts(3_005_844, 0).saturating_mul(c.into()))
 	}
 	fn dispatch_as() -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 5_791_000 picoseconds.
-		Weight::from_parts(6_045_000, 0)
+		// Minimum execution time: 5_556_000 picoseconds.
+		Weight::from_parts(5_909_000, 0)
 	}
 	/// The range of component `c` is `[0, 1000]`.
 	fn force_batch(c: u32, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `0`
 		//  Estimated: `0`
-		// Minimum execution time: 3_858_000 picoseconds.
-		Weight::from_parts(4_025_468, 0)
-			// Standard Error: 1_226
-			.saturating_add(Weight::from_parts(2_777_670, 0).saturating_mul(c.into()))
+		// Minimum execution time: 3_548_000 picoseconds.
+		Weight::from_parts(3_735_851, 0)
+			// Standard Error: 867
+			.saturating_add(Weight::from_parts(2_822_972, 0).saturating_mul(c.into()))
 	}
 }
 
@@ -86,8 +86,7 @@ mod tests {
   use crate::constants::{MAXIMUM_BLOCK_WEIGHT, NORMAL_DISPATCH_RATIO};
   use crate::weights::extrinsic_weights::ExtrinsicBaseWeight;
 
-#[allow(dead_code)]
-struct BlockWeights;
+  struct BlockWeights;
   impl Get<frame_system::limits::BlockWeights> for BlockWeights {
   	fn get() -> frame_system::limits::BlockWeights {
   		frame_system::limits::BlockWeights::builder()
